% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baselearner.R
\name{as.binaryPrediction}
\alias{as.binaryPrediction}
\title{Create a predictive binary result object}
\usage{
as.binaryPrediction(probability, threshold = 0.5)
}
\arguments{
\item{probability}{A vector with probabilities predictions or with bipartitions
prediction for a binary prediction.}

\item{threshold}{A numeric value between 0 and 1 to create the bipartitions.}
}
\value{
An object of type "\code{mlresult}" used by problem transformation
 methods that use binary classifiers. It has only two attributes:
 \code{bipartition} and \code{probability}, that respectively have the
 bipartition and probabilities results.
}
\description{
The transformation methods require a specific data format from base
 classifiers prediction. If you implement a new base method then use this method
 to return the final result of your \code{mlpredict} method.
}
\examples{
# This method is used to implement a mlpredict based method
# In this example we create a random predict method
mlpredict.random <- function (model, newdata, ...) {
   probs <- runif(nrow(newdata), 0, 1)
   as.binaryPrediction(probs)
}

# Use different threshold value
probs <- runif(10, 0, 1)
result <- as.binaryPrediction(probs, 0.6)
}

