% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{as.matrix.mlconfmat}
\alias{as.matrix.mlconfmat}
\title{Convert a multi-label Confusion Matrix to matrix}
\usage{
\method{as.matrix}{mlconfmat}(x, ...)
}
\arguments{
\item{x}{The mlconfmat}

\item{...}{passed to as.matrix}
}
\value{
A confusion matrix with TP, TN, FP and FN columns
}
\description{
Convert a multi-label Confusion Matrix to matrix
}
