% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{as.matrix.mlresult}
\alias{as.matrix.mlresult}
\title{Convert a mlresult to a matrix}
\usage{
\method{as.matrix}{mlresult}(x, ...)
}
\arguments{
\item{x}{The mlresult object}

\item{...}{ignored}
}
\value{
matrix
}
\description{
Convert a mlresult to a matrix
}

