% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlresult.R
\name{as.ranking}
\alias{as.ranking}
\title{Convert a mlresult to a ranking matrix}
\usage{
as.ranking(mlresult, ties.method = "min", ...)
}
\arguments{
\item{mlresult}{The mlresult object}

\item{ties.method}{A character string specifying how ties are treated
(Default: "min"). see \code{\link{rank}} to more details.}

\item{...}{Others parameters passed to the \code{\link{rank}} method.}
}
\value{
matrix with ranking values
}
\description{
Convert a mlresult to a ranking matrix
}
