% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{as.resultMLPrediction}
\alias{as.resultMLPrediction}
\title{Create a predictive multi-label result}
\usage{
as.resultMLPrediction(predictions, probability)
}
\arguments{
\item{predictions}{The list of mlresult obtained from union of all binary
predictions.}

\item{probability}{A logical value. If \code{TRUE} the predicted values are
 the score between 0 and 1, otherwise the values are bipartition 0 or 1.}
}
\value{
A matrix containing the probabilistic values or just predictions.
}
\description{
This function select the correct result and organize them in a
 prediction matrix where the columns are the labels and the rows are the
 test examples. If probability is \code{TRUE} the values contain the labels
 probabilities, otherwise the values are the predictive value "0" or "1".
}
\examples{
...
predictions <- list()
predictions$class1 <- mlpredict(model1, testdata)
predictions$class2 <- mlpredict(model2, testdata)
as.resultMLPrediction(predictions, TRUE)
...
}

