% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{binary.prediction}
\alias{binary.prediction}
\title{Create an object binary.prediction}
\usage{
binary.prediction(bipartition, probability)
}
\arguments{
\item{bipartition}{The classes predictions (bipartition values), only 0 and 1}

\item{probability}{The probability/confidence of a prediction, between 0..1}
}
\value{
An object of type binary.prediction
}
\description{
Create an object binary.prediction
}

