% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{br.predict_model}
\alias{br.predict_model}
\title{Dinamically call the prediction function}
\usage{
br.predict_model(model, newdata, ...)
}
\arguments{
\item{model}{A model generated by \code{\link{br.create_model}} method}

\item{newdata}{The data to be predicted}

\item{...}{Others arguments passed to the base method}
}
\value{
An object of type mlresult generated by the
 \code{\link{as.binaryPrediction}} method
}
\examples{
#Use all models to predict newdata
lapply(models, br.predict_model, newdata = newdata, ...)
}

