% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/br.R
\name{br}
\alias{br}
\title{Binary Relevance for multi-label Classification}
\usage{
br(mdata, base.method = "SVM", ..., save.datasets = FALSE, CORES = 1)
}
\arguments{
\item{mdata}{Object of class \code{\link[mldr]{mldr}}, a multi-label train
dataset (provided by \pkg{mldr} package).}

\item{base.method}{A string with the name of base method. The same base method
  will be used for train all subproblems.

  Default valid options are: \code{'SVM'}, \code{'C4.5'}, \code{'C5.0'},
  \code{'RF'}, \code{'NB'} and \code{'KNN'}. To use other base method see
  \code{\link{mltrain}} and \code{\link{mlpredict}} instructions. (default:
   \code{'SVM'})}

\item{...}{Others arguments passed to the base method for all subproblems}

\item{save.datasets}{Logical indicating whether the binary datasets must be
saved in the model or not. (default: \code{FALSE})}

\item{CORES}{The number of cores to parallelize the training. Values higher
  than 1 require the \pkg{parallel} package. (default: 1)}
}
\value{
An object of class \code{BRmodel} containing the set of fitted
  models, including: \describe{ \item{labels}{A vector with the label names}
  \item{models}{A list of models named by the label names.} \item{datasets}{A
  list of \code{mldBR} named by the label names. Only when the
  \code{save.datasets = TRUE}.} }
}
\description{
Create a Binary Relevance model for multilabel classification.

  Binary Relevance is a simple and effective transformation method to predict
  multi-label data. This is based on the one-versus-all approach to build a
  specific model for each label.
}
\examples{
# Train and predict emotion multilabel dataset using Binary Relevance
library(utiml)
testdata <- emotions$dataset[sample(1:100, 10), emotions$attributesIndexes]

# Use SVM as base method
model <- br(emotions)
pred <- predict(model, testdata)

# Change the default base method and use 4 CORES
model <- br(emotions, "C4.5", CORES = 4)
pred <- predict(model, testdata)

# Set a parameters for all subproblems
model <- br(emotions, "KNN", k=5)
pred <- predict(model, testdata)
}
\references{
Boutell, M. R., Luo, J., Shen, X., & Brown, C. M. (2004). Learning
   multi-label scene classification. Pattern Recognition, 37(9), 1757–1771.
}
\seealso{
Other Transformation methods: \code{\link{brplus}};
  \code{\link{cc}}; \code{\link{dbr}}; \code{\link{ebr}};
  \code{\link{ecc}}; \code{\link{mbr}}; \code{\link{ns}}
}

