% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{check_ensemble_vote}
\alias{check_ensemble_vote}
\title{Verify if a schema vote name is valid}
\usage{
check_ensemble_vote(vote.schema, accept.null = TRUE)
}
\arguments{
\item{vote.schema}{The name of schema vote}

\item{accept.null}{Logical value determine if the vote.schema = NULL is
also valid. (Default: TRUE)}
}
\value{
TRUE or throw an error message otherwise
}
\description{
Verify if a schema vote name is valid
}

