% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{compute_binary_ensemble_votes}
\alias{compute_binary_ensemble_votes}
\title{Compute some vote schema for single-label predictions}
\usage{
compute_binary_ensemble_votes(predictions, vote.schema)
}
\arguments{
\item{predictions}{A list of single-label predictions}

\item{vote.schema}{the vote schema name}
}
\value{
A binary.prediction
}
\description{
Compute some vote schema for single-label predictions
}

