% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{compute_ensemble_votes}
\alias{compute_ensemble_votes}
\title{Generic vote combination for ensemble prediction}
\usage{
compute_ensemble_votes(predictions, method)
}
\arguments{
\item{predictions}{A list of single-label predictions, usually with scores.}

\item{method}{The method that will be compute the results.
This method receive a list of probability values and must return a single
probability value.}
}
\value{
A single-label prediction (a mlprediction object).
}
\description{
Generic vote combination for ensemble prediction
}
\examples{
model <- br(toyml, "KNN")
predictions <- list(
 predict(model, toyml[1:10], k=1)[, "y1"],
 predict(model, toyml[1:10], k=3)[, "y1"],
 predict(model, toyml[1:10], k=5)[, "y1"]
)

## Return the max value for each example
result <- compute_ensemble_votes(predictions, max)
}

