% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{compute_multilabel_ensemble_votes}
\alias{compute_multilabel_ensemble_votes}
\title{Compute the multi-label ensemble predictions based on some vote schema}
\usage{
compute_multilabel_ensemble_votes(predictions, vote.schema,
  probability = TRUE)
}
\arguments{
\item{predictions}{A list of multi-label predictions.}

\item{vote.schema}{Define the way that ensemble must compute the predictions.
The default valid options are:
\describe{
  \code{'avg'}{Compute the proportion of votes, scale data between min and
    max of votes}
  \code{'maj'}{Compute the averages of probabilities},
  \code{'max'}{Compute the votes scaled between 0 and \code{m}
    (number of interations)},
  \code{'min'}{Compute the proportion of votes, scale data between min and
    max of votes}
  \code{'prod'}{Compute the product of all votes for each instance}
}
If NULL then all predictions are returned.}

\item{probability}{A logical value. If \code{TRUE} the predicted values are
the score between 0 and 1, otherwise the values are bipartition 0 or 1.}
}
\value{
A new mlresult that final result
}
\description{
Compute the multi-label ensemble predictions based on some vote schema
}
\note{
You can create your own vote schema, just create a method that receive
 a list with probabilities values and return a single value with the result.
 Your method must be like the methods mean, max and min.
}
\examples{
model <- br(toyml, "KNN")
predictions <- list(
 predict(model, toyml[1:10], k=1),
 predict(model, toyml[1:10], k=3),
 predict(model, toyml[1:10], k=5)
)

result <- compute_multilabel_ensemble_votes(predictions, "avg")
}

