% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threshold.R
\name{compute_subset_correction}
\alias{compute_subset_correction}
\title{Subset Correction of a predicted result}
\usage{
compute_subset_correction(mlresult, train_y, threshold = 0.5)
}
\arguments{
\item{mlresult}{An object of mlresult that contain the scores and bipartition
values.}

\item{train_y}{A matrix/data.frame with all labels values of the training
dataset.}

\item{threshold}{A numeric value between 0 and 1 to use as base to determine
which values needs be reescaled to preserve the corrected labelsets.
(default: 0.5)}
}
\value{
A new mlresult where all results are present in the training
 labelsets.
}
\description{
This method restrict a multi-label learner prediction to only label
combinations whose existence is testified by the (training) data. To this all
labelsets that are predicted but are not found on training data is replaced
by the most similar labelset.
}
\details{
If the most simillar is not unique, those label combinations with higher
frequency in the training data are preferred. The Hamming loss distance is
used to determine the difference between the labelsets.
}
\note{
The original paper describes a method to create only bipartitions, but
 we adapeted the method to use also scores. Based on the threshold values the
 scores higher than the threshold value, but must be lower are changed to
 respect this restriction.
}
\references{
Senge, R., Coz, J. J. del, & Hüllermeier, E. (2013). Rectifying classifier
   chains for multi-label classification. In Workshop of Lernen, Wissen &
   Adaptivität (LWA 2013) (pp. 162–169). Bamberg, Germany.
}
\seealso{
Other threshold: \code{\link{fixed_threshold}}
}

