% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{create_br_model}
\alias{create_br_model}
\title{Create Dynamically the model for Binary Relevance Methods}
\usage{
create_br_model(dataset, ...)
}
\arguments{
\item{dataset}{An object generated by the \code{\link{prepare_br_data}}
method}

\item{...}{Others arguments passed to the base method}
}
\value{
The model dinamically generated
}
\description{
This method is used in the methods to call dinamicaly the base classifier
}

