% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ebr.R
\name{ebr}
\alias{ebr}
\title{Ensemble of Binary Relevance for multi-label Classification}
\usage{
ebr(mdata, base.method = "SVM", m = 10, subsample = 0.75,
  attr.space = 0.5, ..., save.datasets = FALSE, SEED = NULL, CORES = 1)
}
\arguments{
\item{mdata}{Object of class \code{\link[mldr]{mldr}}, a multi-label train
dataset (provided by \pkg{mldr} package).}

\item{base.method}{A string with the name of base method. The same base method
  will be used for train all subproblems.

  Default valid options are: \code{'SVM'}, \code{'C4.5'}, \code{'C5.0'},
  \code{'RF'}, \code{'NB'} and \code{'KNN'}. To use other base method see
  \code{\link{mltrain}} and \code{\link{mlpredict}} instructions. (default:
   \code{'SVM'})}

\item{m}{The number of interations or Binary Relevance models to use in the
ensemble.}

\item{subsample}{A value between 0.1 and 1 to determine the percentage of
training instances must be used for each interation. (default: 0.75)}

\item{attr.space}{A value between 0.1 and 1 to determine the percentage of
attributes must be used for each interation. (default: 0.50)}

\item{...}{Others arguments passed to the base method for all subproblems}

\item{save.datasets}{Logical indicating whether the binary datasets must be
saved in the model or not. (default: \code{FALSE})}

\item{SEED}{A single value, interpreted as an integer to allow obtain the
same results again. (default: \code{NULL})}

\item{CORES}{The number of cores to parallelize the training. Values higher
  than 1 require the \pkg{parallel} package. (default: 1)}
}
\value{
An object of class \code{EBRmodel} containing the set of fitted
  BR models, including: \describe{ \item{rounds}{The number of interations}
  \item{models}{A list of BR models.} \item{nrow}{The number of instances
  used in each training dataset} \item{ncol}{The number of attributes used
  in each training dataset} \item{seed}{The value of the seed, present only
  when the \code{SEED} is defined.}}
}
\description{
Create an Ensemble of Binary Relevance model for
  multilabel classification.

  This model is composed by a set of Binary Relevance models.
  Binary Relevance is a simple and effective transformation method
  to predict multi-label data.
}
\section{Warning}{

   RWeka package does not permit use \code{'C4.5'} in parallel mode, use
   \code{'C5.0'} or \code{'CART'} instead of it.
}
\examples{
# Train and predict emotion multilabel dataset using Ensemble of Binary Relevance
library(utiml)
testdata <- emotions$dataset[sample(1:100, 10), emotions$attributesIndexes]

# Use all default values
model <- ebr(emotions)
pred <- predict(model, testdata)

# Use C4.5 with 100\% of instances and only 5 rounds
model <- ebr(emotions, "C4.5", m = 5, subsample = 1)
pred <- predict(model, testdata)

# Use 75\% of attributes and use a specific seed
model <- ebr(emotions, attr.space = 0.75, SEED = 1)
pred <- predict(model, testdata)
}
\references{
Read, J., Pfahringer, B., Holmes, G., & Frank, E. (2011). Classifier
   chains for multi-label classification. Machine Learning, 85(3), 333–359.

   Read, J., Pfahringer, B., Holmes, G., & Frank, E. (2009).
   Classifier Chains for Multi-label Classification. Machine Learning and
   Knowledge Discovery in Databases, Lecture Notes in Computer Science,
   5782, 254–269.
}
\seealso{
Other Ensemble: \code{\link{ecc}}

Other Transformation methods: \code{\link{brplus}};
  \code{\link{br}}; \code{\link{cc}}; \code{\link{dbr}};
  \code{\link{ecc}}; \code{\link{mbr}}; \code{\link{ns}}
}

