% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{fill_sparce_mldata}
\alias{fill_sparce_mldata}
\title{Fill sparce dataset with 0 or '' values}
\usage{
fill_sparce_mldata(mdata)
}
\arguments{
\item{mdata}{The mldr dataset to be filled.}
}
\value{
a new mldr object.
}
\description{
Transform a sparce dataset filling NA values to 0 or '' based on the column
type. Text columns with numeric values will be modified to numerical.
}
\examples{
sparce.toy <- toyml
sparce.toy$dataset$ratt10[sample(100, 30)] <- NA
complete.toy <- fill_sparce_mldata(sparce.toy)
}
\seealso{
Other pre process: \code{\link{normalize_mldata}},
  \code{\link{remove_attributes}},
  \code{\link{remove_labels}},
  \code{\link{remove_skewness_labels}},
  \code{\link{remove_unique_attributes}},
  \code{\link{remove_unlabeled_instances}},
  \code{\link{replace_nominal_attributes}}
}

