% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_column_of_class}
\alias{get_column_of_class}
\title{Return the column values related with the class}
\usage{
get_column_of_class(data, column, class)
}
\arguments{
\item{data}{A dataset}

\item{column}{The column name or number}

\item{class}{The class value}
}
\value{
A list with the column values that are related with the class value
}
\description{
Return the column values related with the class
}

