% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_dim}
\alias{get_dim}
\title{Return the dimensionality of dataset (attributes / samples)}
\usage{
get_dim(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
numeric
}
\description{
Return the dimensionality of dataset (attributes / samples)
}

