% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_kurtosis}
\alias{get_kurtosis}
\title{Kurtosis mean of dataset}
\usage{
get_kurtosis(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
numeric value of kurtosis
#TODO@references
}
\description{
This is the mean of classes kurtosis, where classes kurtosis are the mean of
attributes kurtosis related with each class. This use only the values related
with each class.
}

