% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_min_max_mean_sd}
\alias{get_min_max_mean_sd}
\title{Calculate the min, max, mean and standart deviation measures for a list}
\usage{
get_min_max_mean_sd(x, name)
}
\arguments{
\item{x}{A list}

\item{name}{The base name}
}
\value{
get_min_max_mean_sd(c(1,2,3,4,5), 'test')
## test_min  test_max test_mean   test_sd
## 1.000000  5.000000  3.000000  1.581139
}
\description{
Calculate the min, max, mean and standart deviation measures for a list
}

