% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_num_att}
\alias{get_num_att}
\title{Return the number of attributes}
\usage{
get_num_att(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
integer
}
\description{
Return the number of attributes
}

