% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_num_classes}
\alias{get_num_classes}
\title{Return the number of classes values}
\usage{
get_num_classes(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
integer
}
\description{
Return the number of classes values
}

