% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_num_samples}
\alias{get_num_samples}
\title{Return the number of samples (instances)}
\usage{
get_num_samples(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
integer
}
\description{
Return the number of samples (instances)
}

