% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_prop_binatt}
\alias{get_prop_binatt}
\title{Return the proportion of binary attributes}
\usage{
get_prop_binatt(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
number between 0 and 1
}
\description{
Binary attributes are those containing two distinct values
}

