% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_prop_posclass}
\alias{get_prop_posclass}
\title{Return the proportion of positive class}
\usage{
get_prop_posclass(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
number between 0 and 1
}
\description{
Return the proportion of positive class
}

