% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_skewness}
\alias{get_skewness}
\title{Skewness mean of a dataset}
\usage{
get_skewness(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
numeric value of skewness
#TODO@references
}
\description{
This is the mean of classes skewness, where classes skewness are the mean of
attributes skewness related with each class. This use only the values related
with each class.
}

