% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_features.R
\name{get_stat_varcoef}
\alias{get_stat_varcoef}
\title{Coefficient of variation}
\usage{
get_stat_varcoef(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
numeric vector with min, max, mean and sd values
}
\description{
Coefficient of variation
}
\references{
Castiello, C., Castellano, G., & Fanelli, A. M. (2005). Meta-data:
 Characterization of Input Features for Meta-learning. Modeling Decisions
 for Artificial Intelligence, 3558, 457–468.
}

