% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prediction.R
\name{is.bipartition}
\alias{is.bipartition}
\title{Test if a mlresult contains crisp values as default}
\usage{
is.bipartition(mlresult)
}
\arguments{
\item{x}{The mlresult object}
}
\value{
logical value
}
\description{
Test if a mlresult contains crisp values as default
}

