% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{is.probability}
\alias{is.probability}
\title{Test if a mlresult contains score values as default}
\usage{
is.probability(mlresult)
}
\arguments{
\item{x}{The mlresult object}
}
\value{
logical value
}
\description{
Test if a mlresult contains score values as default
}

