% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{labelsPhiCorrelationCoefficient}
\alias{labelsPhiCorrelationCoefficient}
\title{Phi Correlation Coefficient}
\usage{
labelsPhiCorrelationCoefficient(mdata)
}
\arguments{
\item{mdata}{Object of class \code{\link[mldr]{mldr}}, a multi-label dataset}
}
\value{
A matrix with all labels phi correlation coefficient. The rows and
columns have the labels and the values are the phi value. The main diagonal
have the 1 value that represents the correlation of a label with itself.
}
\description{
Calculate all labels phi correlation coefficient.
This is a specialized version of the Pearson product moment
correlation coefficient for categorical variables with two
values, also called dichotomous variables.
}
\examples{
library(utiml)
result <- labelsPhiCorrelationCoefficient(emotions)

# Get the phi coefficient between the labels "happy-pleased" and "quiet-still"
result["happy-pleased", "quiet-still"]

# Get all coefficients of a specific label
result[1, ]
}

