% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensemble.R
\name{majority_ensemble_votes}
\alias{majority_ensemble_votes}
\title{Majority vote combination for single-label prediction}
\usage{
majority_ensemble_votes(predictions)
}
\arguments{
\item{predictions}{A list of binary predictions.}
}
\value{
A new single-label prediction.
}
\description{
Compute the single-label prediction using the majority votes schema.
The probabilities result is computed using only the majority instances.
In others words, if a example is predicted as posivite, only the positive
confidences are used to compute the averaged value.
}
\examples{
model <- br(toyml, "KNN")
predictions <- list(
 predict(model, toyml[1:10], k=1)[, "y1"],
 predict(model, toyml[1:10], k=3)[, "y1"],
 predict(model, toyml[1:10], k=5)[, "y1"]
)

result <- majority_ensemble_votes(predictions)
}
\seealso{
Other Ensemble utilites: \code{\link{average_ensemble_votes}};
  \code{\link{maximum_ensemble_votes}};
  \code{\link{minimum_ensemble_votes}};
  \code{\link{product_ensemble_votes}}
}

