% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{merge_mlconfmat}
\alias{merge_mlconfmat}
\title{Join a list of multi-label confusion matrix}
\usage{
merge_mlconfmat(object, ...)
}
\arguments{
\item{object}{A mlconfmat object or a list of mlconfmat objects}

\item{...}{mlconfmat objects}
}
\value{
mlconfmat
}
\description{
Join a list of multi-label confusion matrix
}
