% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{minimum_ensemble_votes}
\alias{minimum_ensemble_votes}
\title{Minimum vote combination for single-label prediction}
\usage{
minimum_ensemble_votes(predictions)
}
\arguments{
\item{predictions}{A list of binary predictions.}
}
\value{
A new single-label prediction.
}
\description{
Compute the single-label prediction using the minimum votes schema. The
probabilities result is computed using the minimum value.
}
\examples{
\dontrun{
model <- br(toyml, "KNN")
predictions <- list(
 predict(model, toyml[1:10], k=1)[, "y1"],
 predict(model, toyml[1:10], k=3)[, "y1"],
 predict(model, toyml[1:10], k=5)[, "y1"]
)

result <- minimum_ensemble_votes(predictions)
}
}
\seealso{
Other Ensemble utilites: \code{\link{average_ensemble_votes}},
  \code{\link{majority_ensemble_votes}},
  \code{\link{maximum_ensemble_votes}},
  \code{\link{product_ensemble_votes}}
}

