% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldr.R
\name{mldata}
\alias{mldata}
\title{Fix the mldr dataset to use factors}
\usage{
mldata(mdata)
}
\arguments{
\item{mdata}{A mldr dataset.}
}
\value{
A mldr object
}
\description{
Fix the mldr dataset to use factors
}
\examples{
toyml <- mldata(toyml)
}

