% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{mldr_getfold}
\alias{mldr_getfold}
\title{Get the multi-labels datasets for k-fold Cross Validation}
\usage{
mldr_getfold(mdata, kfold, n, has.validation = FALSE)
}
\arguments{
\item{mdata}{A dataset of class \code{\link[mldr]{mldr}}.}

\item{kfold}{An object of class \code{mldr_kfolds}, this is obtained
from use of some kfold method.}

\item{n}{The number of fold to separated train and test subsets.}

\item{has.validation}{Logical value that indicate if a validation
dataset will be used. (defaul: \code{FALSE})}
}
\value{
A list contained train and test dataset:
 \describe{
   \code{train}{The mldr dataset with train examples, that inclue all
     examples except those that are in test and validation samples}
   \code{test}{The mldr dataset with test examples, defined by the
     number of the fold}
   \code{validation}{Optionally, only if \code{has.validation = TRUE}.
     The mldr dataset with validation examples}
 }
}
\description{
This is a simple way to use k-fold cross validation.
}
\examples{
library(utiml)
folds <- mldr_random_kfold(emotions, 10)

# Using the first iterationrows <- sample(1:mdata$measures$num.instances)rows <- sample(1:mdata$measures$num.instances)
dataset <- mldr_getfold(emotions, folds, 1)
classifier <- br(dataset$train)
result <- predict(classifier, dataset$test)

# All iterations
for (i in 1:10) {
   dataset <- mldr_getfold(emotions, folds, i)
   #dataset$train
   #dataset$test
}

# Using validation
dataset <- mldr_getfold(emotions, folds, 10, TRUE)
# dataset$train, dataset$test, #dataset$validation
}
\seealso{
Other mldr: \code{\link{mldr_iterative_stratification_holdout}},
  \code{\link{mldr_iterative_stratification_kfold}},
  \code{\link{mldr_random_holdout}},
  \code{\link{mldr_random_kfold}},
  \code{\link{mldr_random_subset}},
  \code{\link{mldr_stratified_holdout}},
  \code{\link{mldr_stratified_kfold}},
  \code{\link{mldr_subset}}

Other sampling: \code{\link{mldr_iterative_stratification_holdout}},
  \code{\link{mldr_iterative_stratification_kfold}},
  \code{\link{mldr_random_holdout}},
  \code{\link{mldr_random_kfold}},
  \code{\link{mldr_stratified_holdout}},
  \code{\link{mldr_stratified_kfold}},
  \code{\link{utiml_holdout}}, \code{\link{utiml_kfold}}
}

