% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mldr.R
\name{mldr_holdout}
\alias{mldr_holdout}
\title{Create distinct partitions of a multi-label dataset}
\usage{
mldr_holdout(mdata, partitions = c(0.7, 0.3), SEED = NULL)
}
\arguments{
\item{mdata}{A dataset of class \code{\link[mldr]{mldr}}.}

\item{partitions}{A list of percentages or a single value.
The sum of all values does not be greater than 1. If a
single value is informed then the complement of them is
applied to generated the second partition. If two or more
values are informed and the sum of them is lower than 1
the partitions will be generated with the ratio informed.
(default: \code{c(0.7, 0.3)})}

\item{SEED}{A single value, interpreted as an integer to allow
 obtain the same results again. (default: \code{NULL})}
}
\value{
A list with at least two datasets sampled as specified
 in partitions parameter.
}
\description{
This method create multi-label dataset for
 train, test, validation or other proposes. The number of
 partitions is defined in \code{partitions} parameter.
 The instances are used in only one partition of divistion.
 Use the SEED parameter to obtain the same result again.
}
\examples{
# Create two partitions with 70\% and 30\% for training and test
datasets <- mldr_holdout(emotions)

# The same result can be obtained as:
datasets <- mldr_holdout(emotions, 0.7)
print(datasets[[1]]$measures)
print(datasets[[2]]$measures)

# Using a SEED and split the dataset in the half
datasets <- mldr_holdout(emotions, 0.5, SEED = 12)

# Split the dataset in three parts
datasets <- mldr_holdout(emotions, c(0.70, 0.15, 0.15))
}
\seealso{
Other mldr: \code{\link{mldr_random_subset}};
  \code{\link{mldr_subset}}
}

