% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{mldr_iterative_stratification_kfold}
\alias{mldr_iterative_stratification_kfold}
\title{Generate labels based stratified k folds for multi-label data}
\usage{
mldr_iterative_stratification_kfold(mdata, k = 10)
}
\arguments{
\item{mdata}{A dataset of class \code{\link[mldr]{mldr}}.}

\item{k}{The number of folds. (default: 10)}
}
\value{
An object of type \code{mldr_kfolds}. This is a list with
 k elements, where each element contains a list of row indexes based
 on the original dataset.
}
\description{
Use this method to generate stratified sampling for multi-labels
  datasets using the iterative stratified algorithm. This method use the
  labels distributions to compute the partitions proportions, however some
  specific label can not occurs in all folds.
}
\examples{
# 2 folds
folds <- mldr_iterative_stratification_kfold(emotions, 2)

# 10 folds
folds <- mldr_iterative_stratification_kfold(emotions, 10)
}
\references{
Sechidis, K., Tsoumakas, G., & Vlahavas, I. (2011). On the
 stratification of multi-label data.If partitions have names, they are used to name the return. In Proceedings of the Machine
 Learningand Knowledge Discovery in Databases - European Conference,
 ECML PKDD (pp. 145–158).
}
\seealso{
\code{\link{mldr_getfold}}

Other mldr: \code{\link{mldr_getfold}};
  \code{\link{mldr_iterative_stratification_holdout}};
  \code{\link{mldr_random_holdout}};
  \code{\link{mldr_random_kfold}};
  \code{\link{mldr_random_subset}};
  \code{\link{mldr_stratified_holdout}};
  \code{\link{mldr_stratified_kfold}};
  \code{\link{mldr_subset}}

Other sampling: \code{\link{mldr_getfold}};
  \code{\link{mldr_iterative_stratification_holdout}};
  \code{\link{mldr_random_holdout}};
  \code{\link{mldr_random_kfold}};
  \code{\link{mldr_stratified_holdout}};
  \code{\link{mldr_stratified_kfold}};
  \code{\link{utiml_holdout}}; \code{\link{utiml_kfold}}
}

