% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mldr_normalize}
\alias{mldr_normalize}
\title{Normalize dataset attributes}
\usage{
mldr_normalize(mdata)
}
\arguments{
\item{mdata}{The mldr dataset to be normalized.}
}
\value{
a new mldr object.
}
\description{
Normalize all numerical attributes to values between
 0 and 1.
}
\examples{
mldr_normalize(emotions)
}

