% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{mldr_random_holdout}
\alias{mldr_random_holdout}
\title{Create distinct partitions of a multi-label dataset}
\usage{
mldr_random_holdout(mdata, partitions = c(0.7, 0.3))
}
\arguments{
\item{mdata}{A dataset of class \code{\link[mldr]{mldr}}.}

\item{partitions}{A list of percentages or a single value.
 The sum of all values does not be greater than 1. If a
 single value is informed then the complement of them is
 applied to generated the second partition. If two or more
 values are informed and the sum of them is lower than 1
 the partitions will be generated with the informed proportion.
 If partitions have names, they are used to name the return.
 (default: \code{c(0.7, 0.3)})}
}
\value{
A list with at least two datasets sampled as specified
 in partitions parameter.
}
\description{
This method creates multi-label dataset for
 train, test, validation or other proposes. The number of
 partitions is defined in \code{partitions} parameter.
 The instances are used in only one partition of divistion.
}
\examples{
# Create two partitions with 70\% and 30\% for training and test
datasets <- mldr_random_holdout(emotions)

# The same result can be obtained as:
datasets <- mldr_random_holdout(emotions, 0.7)
print(datasets[[1]]$measures)
print(datasets[[2]]$measures)

# Split the dataset in three parts
partitions <- c("train" = 0.70, "validation" = 0.15, "test" = 0.15)
dataset <- mldr_random_holdout(emotions, partitions)
print(dataset$train)
print(dataset$validation)
print(dataset$test)
}
\seealso{
Other mldr: \code{\link{mldr_getfold}};
  \code{\link{mldr_iterative_stratification_holdout}};
  \code{\link{mldr_iterative_stratification_kfold}};
  \code{\link{mldr_random_kfold}};
  \code{\link{mldr_random_subset}};
  \code{\link{mldr_stratified_holdout}};
  \code{\link{mldr_stratified_kfold}};
  \code{\link{mldr_subset}}

Other sampling: \code{\link{mldr_getfold}};
  \code{\link{mldr_iterative_stratification_holdout}};
  \code{\link{mldr_iterative_stratification_kfold}};
  \code{\link{mldr_random_kfold}};
  \code{\link{mldr_stratified_holdout}};
  \code{\link{mldr_stratified_kfold}};
  \code{\link{utiml_holdout}}; \code{\link{utiml_kfold}}
}

