% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampling.R
\name{mldr_random_kfold}
\alias{mldr_random_kfold}
\title{Generate random k folds for multi-label data}
\usage{
mldr_random_kfold(mdata, k = 10)
}
\arguments{
\item{mdata}{A dataset of class \code{\link[mldr]{mldr}}.}

\item{k}{The number of folds. (default: 10)}
}
\value{
An object of type \code{mldr_kfolds}. This is a list with
 k elements, where each element contains a list of row indexes based
 on the original dataset.
}
\description{
Use this method to generate random sampling for multi-labels
  datasets. This may generate folds with differents proportions of labels.
}
\examples{
# 2 folds
folds <- mldr_random_kfold(emotions, 2)

# 10 folds
folds <- mldr_random_kfold(emotions, 10)
}
\seealso{
\code{\link{mldr_getfold}}

Other mldr: \code{\link{mldr_getfold}};
  \code{\link{mldr_iterative_stratification_holdout}};
  \code{\link{mldr_iterative_stratification_kfold}};
  \code{\link{mldr_random_holdout}};
  \code{\link{mldr_random_subset}};
  \code{\link{mldr_stratified_holdout}};
  \code{\link{mldr_stratified_kfold}};
  \code{\link{mldr_subset}}

Other sampling: \code{\link{mldr_getfold}};
  \code{\link{mldr_iterative_stratification_holdout}};
  \code{\link{mldr_iterative_stratification_kfold}};
  \code{\link{mldr_random_holdout}};
  \code{\link{mldr_stratified_holdout}};
  \code{\link{mldr_stratified_kfold}};
  \code{\link{utiml_holdout}}; \code{\link{utiml_kfold}}
}

