% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mldr.R
\name{mldr_random_subset}
\alias{mldr_random_subset}
\title{Create a random subset of a dataset}
\usage{
mldr_random_subset(mdata, num.rows, num.cols)
}
\arguments{
\item{mdata}{A \code{mldr} objetc dataset}

\item{num.rows}{The number of expected rows (instances)}

\item{num.cols}{The number of expected columns (attributes)}
}
\value{
A new mldr subset
}
\description{
Create a random subset of a dataset
}
\examples{
#Return a emotion subset with 100 examples and 40 predictive attributes
mldr_random_subset(emotions, 100, 40)
}
\seealso{
Other mldr: \code{\link{mldr_getfold}};
  \code{\link{mldr_iterative_stratification_holdout}};
  \code{\link{mldr_iterative_stratification_kfold}};
  \code{\link{mldr_random_holdout}};
  \code{\link{mldr_random_kfold}};
  \code{\link{mldr_stratified_holdout}};
  \code{\link{mldr_stratified_kfold}};
  \code{\link{mldr_subset}}
}

