% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mldr_remove_labels}
\alias{mldr_remove_labels}
\title{Remove unusual or very common labels}
\usage{
mldr_remove_labels(mdata, t = 1)
}
\arguments{
\item{mdata}{The mldr dataset to remove.}

\item{t}{Threshold value.}
}
\value{
a new mldr object.
}
\description{
Remove the labels that have smaller or higher examples based on
 a specific threshold value.
}
\examples{
mldr_remove_labels(emotions)
}

