% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mldr_remove_unique_attributes}
\alias{mldr_remove_unique_attributes}
\title{Remove unique attributes}
\usage{
mldr_remove_unique_attributes(mdata)
}
\arguments{
\item{mdata}{The mldr dataset to remove.}
}
\value{
a new mldr object.
}
\description{
Remove the attributes that have a single value.
 Observation: NA is considered a value.
}
\examples{
mldr_remove_unique_attributes(emotions)
}

