% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mldr_remove_unlabeled_instances}
\alias{mldr_remove_unlabeled_instances}
\title{Remove examples without labels}
\usage{
mldr_remove_unlabeled_instances(mdata)
}
\arguments{
\item{mdata}{The mldr dataset to remove.}
}
\value{
a new mldr object.
}
\description{
Remove the examples that there are not labels.
}
\examples{
mldr_remove_unlabeled_instances(emotions)
}

