% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mldr_replace_nominal_attributes}
\alias{mldr_replace_nominal_attributes}
\title{Replace nominal attributes}
\usage{
mldr_replace_nominal_attributes(mdata, ordinal.attributes = list())
}
\arguments{
\item{mdata}{The mldr dataset to remove.}

\item{ordinal.attributes}{Not yet, but it will be used to specify which attributes need to be replaced.}
}
\value{
a new mldr object.
}
\description{
Replace the nominal attributes by binary attributes.
}
\examples{
mldr_replace_nominal_attributes(emotions)
}

