% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{mldr_subset}
\alias{mldr_subset}
\title{Create a subset of a dataset}
\usage{
mldr_subset(mdata, rows, cols)
}
\arguments{
\item{mdata}{A \code{mldr} objetc dataset}

\item{rows}{A vector with the row indexes (instance indexes)}

\item{cols}{A vector with the column indexes (attribute indexes)}
}
\value{
A new mldr subset
}
\description{
Create a subset of a dataset
}
\examples{
#Return a emotion subset with the first 40 examples and only with two attributes
rows <- 1:40
cols <- c("Mean_Acc1298_Mean_Mem40_Centroid", "Mean_Acc1298_Mean_Mem40_MFCC_10")
mldr_subset(emotions, rows, cols)
}

