% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_accuracy}
\alias{mlmeasure_accuracy}
\title{MULTILABEL MEASURES -------------------------------------------------------
Multi-label Accuracy Measure}
\usage{
mlmeasure_accuracy(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}

\item{...}{ignored}
}
\description{
MULTILABEL MEASURES -------------------------------------------------------
Multi-label Accuracy Measure
}
\references{
Gibaja, E., & Ventura, S. (2015). A Tutorial on Multilabel
 Learning. ACM Comput. Surv., 47(3), 52:1-52:38.
}

