% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_binary_AUC}
\alias{mlmeasure_binary_AUC}
\title{Compute the binary AUC}
\usage{
mlmeasure_binary_AUC(scores, labels)
}
\arguments{
\item{scores}{The probability/score from a single label}

\item{labels}{The expected label predictions}
}
\value{
AUC value between 0 and 1
}
\description{
Compute the binary AUC
}

