% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_binary_f1}
\alias{mlmeasure_binary_f1}
\title{Compute the binary F1 measure}
\usage{
mlmeasure_binary_f1(TP, FP, TN, FN)
}
\arguments{
\item{TP}{The number of True Positive values}

\item{FP}{The number of False Positive values}

\item{TN}{The number of True Negative values}

\item{FN}{The number of False Negative values}
}
\value{
F1 measure value between 0 and 1
}
\description{
Compute the binary F1 measure
}

