% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_hamming_loss}
\alias{mlmeasure_hamming_loss}
\title{Multi-label Hamming Loss Measure}
\usage{
mlmeasure_hamming_loss(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}

\item{...}{ignored}
}
\description{
Multi-label Hamming Loss Measure
}
\references{
Schapire, R. E., & Singer, Y. (1999). Improved boosting
 algorithm using confidence-rated predictions. Machine Learning, 297-336.
}

