% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_macro_AUC}
\alias{mlmeasure_macro_AUC}
\title{Multi-label Macro-AUC Measure}
\usage{
mlmeasure_macro_AUC(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}
}
\description{
Multi-label Macro-AUC Measure
}
\references{
Zhang, M.-L., & Zhou, Z.-H. (2014). A Review on Multi-Label
 Learning Algorithms. IEEE Transactions on Knowledge and Data Engineering,
 26(8), 1819–1837.
}

