% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_macro_f1}
\alias{mlmeasure_macro_f1}
\title{Multi-label Macro-F1 Measure}
\usage{
mlmeasure_macro_f1(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}

\item{...}{ignored}
}
\description{
Multi-label Macro-F1 Measure
}
\references{
Gibaja, E., & Ventura, S. (2015). A Tutorial on Multilabel
 Learning. ACM Comput. Surv., 47(3), 52:1-52:38.
}

