% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_margin_loss}
\alias{mlmeasure_margin_loss}
\title{Multi-label Margin Loss Measure}
\usage{
mlmeasure_margin_loss(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}
}
\description{
Multi-label Margin Loss Measure
}
\references{
Loza Mencía, E., & Fürnkranz, J. (2010). Efficient Multilabel
Classification Algorithms for Large-Scale Problems in the Legal Domain.
In Semantic Processing of Legal Texts (pp. 192–215).
}

