% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_one_error}
\alias{mlmeasure_one_error}
\title{Multi-label One Error Measure}
\usage{
mlmeasure_one_error(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}
}
\description{
Multi-label One Error Measure
}
\references{
Schapire, R. E., & Singer, Y. (2000). BoosTexter: A boosting-
based system for text categorization. Machine Learning, 39(2), 135–168.
}

