% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_recall}
\alias{mlmeasure_recall}
\title{Multi-label Recall Measure}
\usage{
mlmeasure_recall(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}
}
\description{
Multi-label Recall Measure
}
\references{
Godbole, S., & Sarawagi, S. (2004). Discriminative Methods for
Multi-labeled Classification. In Proceedings of the 8th Pacific-Asia
Conference on Knowledge Discovery and Data Mining (PAKDD 2004) (pp. 22–30).
}

