% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_subset_accuracy}
\alias{mlmeasure_subset_accuracy}
\title{Multi-label Subset Accuracy Measure}
\usage{
mlmeasure_subset_accuracy(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}
}
\description{
Multi-label Subset Accuracy Measure
}
\references{
Zhu, S., Ji, X., Xu, W., & Gong, Y. (2005). Multilabelled
 Classification Using Maximum Entropy Method. In Proceedings of the 28th
 Annual International ACM SIGIR Conference on Research and Development in
 Information Retrieval (SIGIR’05) (pp. 274–281).
}

