% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{mlresult}
\alias{mlresult}
\title{Create an object mlresult}
\usage{
mlresult(bipartition, probability)
}
\arguments{
\item{bipartition}{The classes predictions (bipartition values), only 0 and 1}

\item{probability}{The probability/confidence of a prediction, between 0..1}
}
\value{
An object of type mlresult
}
\description{
Create an object mlresult
}

