% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(real, predicted)
}
\arguments{
\item{real}{A list with the expected values}

\item{predicted}{A list with the predicted values}
}
\value{
The calculed mean squared error
}
\description{
Mean Squared Error
}

