% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{multilabel.prediction}
\alias{multilabel.prediction}
\title{Create an object mlresult}
\usage{
multilabel.prediction(bipartitions, probabilities, probability)
}
\arguments{
\item{bipartitions}{The matrix of predictions (bipartition values), only 0 and 1}

\item{probabilities}{The matrix of probability/confidence of a prediction, between 0..1}

\item{probability}{A logical value. If \code{TRUE} the predicted values are
 the score between 0 and 1, otherwise the values are bipartition 0 or 1.}
}
\value{
An object of type mlresult
}
\description{
Create an object mlresult
}

