% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{multilabel_accuracy}
\alias{multilabel_accuracy}
\title{MULTILABEL MEASURES -------------------------------------------------------}
\usage{
multilabel_accuracy(mlconfmat, ...)
}
\description{
MULTILABEL MEASURES -------------------------------------------------------
}

