% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{multilabel_binary_accuracy}
\alias{multilabel_binary_accuracy}
\title{BINARY MEASURES -----------------------------------------------------------
Compute the binary accuracy}
\usage{
multilabel_binary_accuracy(TP, FP, TN, FN)
}
\arguments{
\item{TP}{The number of True Positive values}

\item{FP}{The number of False Positive values}

\item{TN}{The number of True Negative values}

\item{FN}{The number of False Negative values}
}
\value{
Accuracy value between 0 and 1
}
\description{
BINARY MEASURES -----------------------------------------------------------
Compute the binary accuracy
}

