% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{multilabel_binary_precision}
\alias{multilabel_binary_precision}
\title{Compute the binary precision}
\usage{
multilabel_binary_precision(TP, FP, TN, FN)
}
\arguments{
\item{TP}{The number of True Positive values}

\item{FP}{The number of False Positive values}

\item{TN}{The number of True Negative values}

\item{FN}{The number of False Negative values}
}
\value{
Precision value between 0 and 1
}
\description{
Compute the binary precision
}

