% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measures.R
\name{multilabel_loss_function}
\alias{multilabel_loss_function}
\title{Cost-based loss function for multi-label classification}
\usage{
multilabel_loss_function(mdata, mlresult, cost = 0.5)
}
\arguments{
\item{mdata}{A mldr dataset containing the test data.}

\item{mlresult}{An object of mlresult that contain the scores and bipartition
values.}

\item{cost}{The cost of classification each positive label. If a single value
is informed then the all labels have tha same cost.}
}
\description{
Cost-based loss function for multi-label classification
}
\references{
Al-Otaibi, R., Flach, P., & Kull, M. (2014). Multi-label Classification: A
 Comparative Study on Threshold Selection Methods. In First International
 Workshop on Learning over Multiple Contexts (LMCE) at ECML-PKDD 2014.
}

