% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{multilabel_measures}
\alias{multilabel_measures}
\title{Return the name of all measures}
\usage{
multilabel_measures()
}
\value{
array of character contained the measures names.
}
\description{
Return the name of all measures
}
\examples{
multilabel_measures()
}
\seealso{
Other evaluation: \code{\link{cv}},
  \code{\link{multilabel_confusion_matrix}},
  \code{\link{multilabel_evaluate}}
}
