% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{normalize_mldata}
\alias{normalize_mldata}
\title{Normalize numerical attributes}
\usage{
normalize_mldata(mdata)
}
\arguments{
\item{mdata}{The mldr dataset to be normalized.}
}
\value{
a new mldr object.
}
\description{
Normalize all numerical attributes to values between 0 and 1. The highest
value is changed to 1 and the lowest value to 0.
}
\examples{
norm.toy <- normalize_mldata(toyml)
}
\seealso{
Other pre process: \code{\link{fill_sparce_mldata}},
  \code{\link{remove_attributes}},
  \code{\link{remove_labels}},
  \code{\link{remove_skewness_labels}},
  \code{\link{remove_unique_attributes}},
  \code{\link{remove_unlabeled_instances}},
  \code{\link{replace_nominal_attributes}}
}

