% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/threshold.R
\name{ns.subsetcorrection}
\alias{ns.subsetcorrection}
\title{Subset Correction of a predicted result}
\usage{
ns.subsetcorrection(predicted_y, train_y)
}
\arguments{
\item{predicted_y}{A matrix of bipartitions predictions, where the columns
represent the labels values and the rows the examples values.}

\item{train_y}{A matrix with all labels values of the training data.}
}
\value{
A new matrix of bipartitions where all results are present in the
 training labelsets.
}
\description{
This method restrict a multi-label learner prediction to only
 label combinations whose existence is testified by the (training) data. To
 this all labelsets that are predicted but are not found on training data is
 replaced by the most similar labelset.

 If the most simillar is not unique, those label combinations with higher
 frequency in the training data are preferred. The Hamming loss is used
 to determine the difference between the labelsets.
}
\examples{
library(utiml)

trainls <- matrix(c(1,1,1, 1,0,1, 1,0,1, 1,0,0, 1,1,1, 1,0,1), ncol = 3, byrow = TRUE)
colnames(trainls) <- c("c1", "c2", "c3")

predict <- matrix(c(1,1,0, 0,0,1, 0,0,0), ncol = 3, byrow = TRUE)
colnames(predict) <- c("c1", "c2", "c3")

ns.subsetcorrection(predict, trainls)
#       c2 c1 c3
# [1,]  1  1  1
# [2,]  1  0  1
# [3,]  1  0  0
}
\references{
Senge, R., Coz, J. J. del, & Hüllermeier, E. (2013). Rectifying classifier
   chains for multi-label classification. In Workshop of Lernen, Wissen &
   Adaptivität (LWA 2013) (pp. 162–169). Bamberg, Germany.
}
\seealso{
\code{\link[=ns.subsetcorrection.score]{Subset score correction}}

Other threshold: \code{\link{ns.subsetcorrection.score}};
  \code{\link{simple.threshold}}
}

