% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threshold.R
\name{ns.subsetcorrection.score}
\alias{ns.subsetcorrection.score}
\title{Subset Score Correction of a predicted result}
\usage{
ns.subsetcorrection.score(predicted_y, train_y, threshold = 0.5)
}
\arguments{
\item{predicted_y}{A matrix of probabilites/scores predictions, where the
columns represent the labels values and the rows the examples values.}

\item{train_y}{A matrix with all labels values of the training data.}

\item{threshold}{A numeric value between 0 and 1 to use as base to determine
 which values needs be reescaled to preserve the corrected labelsets.
 (default: 0.5)}
}
\value{
A new matrix of probabilities where all results are present in the
 training labelsets, based on the threshold value.
}
\description{
This method is based on \code{\link{ns.subsetcorrection}},
 however, instead of using bipartition values, this use the probabilities
 scores.

 As default method does not specify how to do this, this method changes the
 values of all labels scores that based on a threshold value does not in the
 correct set. The values are distributed between the superior and inferior
 nearest value from threshold, limiting it to 0.05 from threshold value.
}
\examples{
library(utiml)

trainls <- matrix(c(1,1,1, 1,0,1, 1,0,1, 1,0,0, 1,1,1, 1,0,1), ncol = 3, byrow = TRUE)
colnames(trainls) <- c("c1", "c2", "c3")

predict <- matrix(c(
   0.57, 0.84, 0.27,
   0.40, 0.49, 0.74,
   0.39, 0.62, 0.45
 ), ncol = 3, byrow = TRUE)
colnames(predict) <- c("c1", "c2", "c3")

ns.subsetcorrection.score(predict, trainls)
#       c2    c1   c3
# [1,]  0.570 0.84 0.5189
# [2,]  0.540 0.49 0.7400
# [3,]  0.539 0.62 0.5450
}
\seealso{
\code{\link[=ns.subsetcorrection]{Subset correction}}

Other threshold: \code{\link{ns.subsetcorrection}};
  \code{\link{simple.threshold}}
}

