% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{+.mlconfmat}
\alias{+.mlconfmat}
\title{Join two multi-label confusion matrix}
\usage{
\method{+}{mlconfmat}(mlcm1, mlcm2)
}
\arguments{
\item{mlcm1}{A mlconfmat}

\item{mlcm2}{Other mlconfmat}
}
\value{
mlconfmat
}
\description{
Join two multi-label confusion matrix
}
