% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_cc.R
\name{predict.CCmodel}
\alias{predict.CCmodel}
\title{Predict Method for Classifier Chains}
\usage{
\method{predict}{CCmodel}(object, newdata,
  probability = getOption("utiml.use.probs", TRUE), ...)
}
\arguments{
\item{object}{Object of class '\code{CCmodel}'.}

\item{newdata}{An object containing the new input data. This must be a
matrix, data.frame or a mldr object.}

\item{probability}{Logical indicating whether class probabilities should be
returned. (Default: \code{getOption("utiml.use.probs", TRUE)})}

\item{...}{Others arguments passed to the base method prediction for all
subproblems.}
}
\value{
An object of type mlresult, based on the parameter probability.
}
\description{
This function predicts values based upon a model trained by \code{cc}.
}
\note{
The Classifier Chains prediction can not be parellelized.
}
\examples{
\dontrun{
# Predict SVM scores
model <- cc(toyml)
pred <- predict(model, toyml)

# Predict SVM bipartitions
pred <- predict(model, toyml, prob = FALSE)

# Passing a specif parameter for SVM predict method
pred <- predict(model, toyml, na.action = na.fail)
}
}
\seealso{
\code{\link[=cc]{Classifier Chains (CC)}}
}

