% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_ebr.R
\name{predict.EBRmodel}
\alias{predict.EBRmodel}
\title{Predict Method for Ensemble of Binary Relevance}
\usage{
\method{predict}{EBRmodel}(object, newdata, vote.schema = "maj",
  probability = getOption("utiml.use.probs", TRUE), ...,
  CORES = getOption("utiml.cores", 1))
}
\arguments{
\item{object}{Object of class '\code{EBRmodel}'.}

\item{newdata}{An object containing the new input data. This must be a
matrix, data.frame or a mldr object.}

\item{vote.schema}{Define the way that ensemble must compute the predictions.
The default valid options are:
\describe{
  \code{'avg'}{Compute the proportion of votes, scale data between min and
    max of votes}
  \code{'maj'}{Compute the averages of probabilities},
  \code{'max'}{Compute the votes scaled between 0 and \code{m}
    (number of interations)},
  \code{'min'}{Compute the proportion of votes, scale data between min and
    max of votes}
  \code{'prod'}{Compute the product of all votes for each instance}
}
If \code{NULL} then all predictions are returned. (Default: \code{'maj'})}

\item{probability}{Logical indicating whether class probabilities should be
returned. (Default: \code{getOption("utiml.use.probs", TRUE)})}

\item{...}{Others arguments passed to the base method prediction for all
subproblems.}

\item{CORES}{The number of cores to parallelize the training. Values higher
than 1 require the \pkg{parallel} package. (Default:
\code{options("utiml.cores", 1)})}
}
\value{
An object of type mlresult, based on the parameter probability.
}
\description{
This method predicts values based upon a model trained by \code{\link{ebr}}.
}
\examples{
\dontrun{
# Predict SVM scores
model <- ebr(toyml)
pred <- predict(model, toyml)

# Predict SVM bipartitions running in 6 cores
pred <- predict(model, toyml, prob = FALSE, CORES = 6)

# Return the classes with the highest score
pred <- predict(model, toyml, vote = 'max')
}
}
\seealso{
\code{\link[=ebr]{Ensemble of Binary Relevance (EBR)}}
}

