% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ns.R
\name{predict.NSmodel}
\alias{predict.NSmodel}
\title{Predict Method for Nested Stacking}
\usage{
\method{predict}{NSmodel}(object, newdata, probability = TRUE, ...)
}
\arguments{
\item{object}{Object of class "\code{NSmodel}", created by \code{\link{ns}} method.}

\item{newdata}{An object containing the new input data. This must be a matrix or
data.frame object containing the same size of training data or a mldr object.}

\item{probability}{Logical indicating whether class probabilities should be returned.
  (default: \code{TRUE})}

\item{...}{Others arguments passed to the base method prediction for all
subproblems.}
}
\value{
A matrix containing the probabilistic values or just predictions (only when
  \code{probability = FALSE}). The rows indicate the predicted object and the
  columns indicate the labels.
}
\description{
This function predicts values based upon a model trained by \code{ns}.
 The scores of the prediction was adapted once this method uses a correction of
 labelsets to predict only classes present on training data. To more information
 about this implementation see \code{\link{ns.subsetcorrection.score}}.
}
\examples{
# Emotion multi-label dataset using Nested Stacking
dataset <- mldr_random_holdout(emotions, c(train=0.9, test=0.1))

# Predict SVM scores
model <- ns(dataset$train)
pred <- predict(model, dataset$test)

# Predict SVM bipartitions
pred <- predict(model, dataset$test, probability = FALSE)

# Passing a specif parameter for SVM predict method
pred <- predict(model, dataset$test, na.action = na.fail)
}
\seealso{
\code{\link[=ns]{Nested Stacking (NS)}}
}

