% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_prudent.R
\name{predict.PruDentmodel}
\alias{predict.PruDentmodel}
\title{Predict Method for PruDent}
\usage{
\method{predict}{PruDentmodel}(object, newdata,
  probability = getOption("utiml.use.probs", TRUE), ...,
  CORES = getOption("utiml.cores", 1))
}
\arguments{
\item{object}{Object of class '\code{PruDentmodel}'.}

\item{newdata}{An object containing the new input data. This must be a
matrix, data.frame or a mldr object.}

\item{probability}{Logical indicating whether class probabilities should be
returned. (Default: \code{getOption("utiml.use.probs", TRUE)})}

\item{...}{Others arguments passed to the base method prediction for all
subproblems.}

\item{CORES}{The number of cores to parallelize the training. Values higher
than 1 require the \pkg{parallel} package. (Default:
\code{options("utiml.cores", 1)})}
}
\value{
An object of type mlresult, based on the parameter probability.
}
\description{
This function predicts values based upon a model trained by \code{prudent}.
}
\examples{
\dontrun{
# Predict SVM scores
model <- prudent(toyml)
pred <- predict(model, toyml)

# Predict SVM bipartitions
pred <- predict(model, toyml, probability = FALSE)

# Passing a specif parameter for SVM predict method
pred <- predict(model, toyml, na.action = na.fail)
}
}
\seealso{
\code{\link[=prudent]{PruDent}}
}

