% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_br_model}
\alias{predict_br_model}
\title{Dinamically call the prediction function}
\usage{
predict_br_model(model, newdata, ...)
}
\arguments{
\item{model}{A model generated by \code{\link{create_br_model}} method}

\item{newdata}{The data to be predicted}

\item{...}{Others arguments passed to the base method}
}
\value{
An object of type mlresult generated by the
 \code{\link{as.binaryPrediction}} method
}
\description{
Dinamically call the prediction function
}
\examples{
#Use all models to predict newdata
lapply(models, predict_br_model, newdata = newdata, ...)
}

