% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prepare_br_data}
\alias{prepare_br_data}
\title{Prepare a Binary MultiLabel Data to be processed}
\usage{
prepare_br_data(dataset, classname, base.method, ...)
}
\arguments{
\item{dataset}{A data.frame with the data (the last column must be the class
column)}

\item{classname}{The name of specific class of the object}

\item{base.method}{The name of the base method that will process this dataset}

\item{...}{Extra parameters for adding in the dataset object}
}
\value{
A list with data, labelname, labelindex and methodname.
   This list has three classes: mltransformation, baseMETHODNAME and
   a specific name
}
\description{
Prepare a Binary MultiLabel Data to be processed
}
\examples{
# ...
tbl <- prepare_br_data(dataframe, 'mldBR', 'SVM')
# ...
}

