% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_br.R
\name{print.BRmodel}
\alias{print.BRmodel}
\title{Print BR model}
\usage{
\method{print}{BRmodel}(x, ...)
}
\arguments{
\item{x}{The br model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print BR model
}
