% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_brplus.R
\name{print.BRPmodel}
\alias{print.BRPmodel}
\title{Print BRP model}
\usage{
\method{print}{BRPmodel}(x, ...)
}
\arguments{
\item{x}{The brp model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print BRP model
}
