% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_cc.R
\name{print.CCmodel}
\alias{print.CCmodel}
\title{Print CC model}
\usage{
\method{print}{CCmodel}(x, ...)
}
\arguments{
\item{x}{The cc model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print CC model
}
