% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_clr.R
\name{print.CLRmodel}
\alias{print.CLRmodel}
\title{Print CLR model}
\usage{
\method{print}{CLRmodel}(x, ...)
}
\arguments{
\item{x}{The br model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print CLR model
}
