% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_ctrl.R
\name{print.CTRLmodel}
\alias{print.CTRLmodel}
\title{Print CTRL model}
\usage{
\method{print}{CTRLmodel}(x, ...)
}
\arguments{
\item{x}{The ctrlmodel}

\item{...}{ignored}
}
\description{
Print CTRL model
}
