% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_dbr.R
\name{print.DBRmodel}
\alias{print.DBRmodel}
\title{Print DBR model}
\usage{
\method{print}{DBRmodel}(x, ...)
}
\arguments{
\item{x}{The dbr model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print DBR model
}
