% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_ebr.R
\name{print.EBRmodel}
\alias{print.EBRmodel}
\title{Print EBR model}
\usage{
\method{print}{EBRmodel}(x, ...)
}
\arguments{
\item{x}{The ebr model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print EBR model
}
