% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_ecc.R
\name{print.ECCmodel}
\alias{print.ECCmodel}
\title{Print ECC model}
\usage{
\method{print}{ECCmodel}(x, ...)
}
\arguments{
\item{x}{The ecc model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print ECC model
}
