% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_esl.R
\name{print.ESLmodel}
\alias{print.ESLmodel}
\title{Print ESL model}
\usage{
\method{print}{ESLmodel}(x, ...)
}
\arguments{
\item{x}{The esl model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print ESL model
}
