% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{print.kFoldPartition}
\alias{print.kFoldPartition}
\title{Print a kFoldPartition object}
\usage{
\method{print}{kFoldPartition}(x, ...)
}
\arguments{
\item{x}{The kFoldPartition object}

\item{...}{ignored}
}
\value{
No return value, called for print folds' detail
}
\description{
Print a kFoldPartition object
}
