% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_lp.R
\name{print.LPmodel}
\alias{print.LPmodel}
\title{Print LP model}
\usage{
\method{print}{LPmodel}(x, ...)
}
\arguments{
\item{x}{The lp model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print LP model
}
