% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{print.majorityModel}
\alias{print.majorityModel}
\title{Print Majority model}
\usage{
\method{print}{majorityModel}(x, ...)
}
\arguments{
\item{x}{The base model}

\item{...}{ignored}
}
\description{
Print Majority model
}
