% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_mbr.R
\name{print.MBRmodel}
\alias{print.MBRmodel}
\title{Print MBR model}
\usage{
\method{print}{MBRmodel}(x, ...)
}
\arguments{
\item{x}{The mbr model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print MBR model
}
