% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{print.mlconfmat}
\alias{print.mlconfmat}
\title{Print a Multi-label Confusion Matrix}
\usage{
\method{print}{mlconfmat}(x, ...)
}
\arguments{
\item{x}{The mlconfmat}

\item{...}{ignored}
}
\value{
No return value, called for print a confusion matrix
}
\description{
Print a Multi-label Confusion Matrix
}
