% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_mlknn.R
\name{print.MLKNNmodel}
\alias{print.MLKNNmodel}
\title{Print MLKNN model}
\usage{
\method{print}{MLKNNmodel}(x, ...)
}
\arguments{
\item{x}{The mlknn model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print MLKNN model
}
