% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_ns.R
\name{print.NSmodel}
\alias{print.NSmodel}
\title{Print NS model}
\usage{
\method{print}{NSmodel}(x, ...)
}
\arguments{
\item{x}{The ns model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print NS model
}
