% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_prudent.R
\name{print.PruDentmodel}
\alias{print.PruDentmodel}
\title{Print PruDent model}
\usage{
\method{print}{PruDentmodel}(x, ...)
}
\arguments{
\item{x}{The prudent model}

\item{...}{ignored}
}
\description{
Print PruDent model
}
