% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_ps.R
\name{print.PSmodel}
\alias{print.PSmodel}
\title{Print PS model}
\usage{
\method{print}{PSmodel}(x, ...)
}
\arguments{
\item{x}{The ps model}

\item{...}{ignored}
}
\value{
No return value, called for print model's detail
}
\description{
Print PS model
}
