% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_rdbr.R
\name{print.RDBRmodel}
\alias{print.RDBRmodel}
\title{Print RDBR model}
\usage{
\method{print}{RDBRmodel}(x, ...)
}
\arguments{
\item{x}{The rdbr model}

\item{...}{ignored}
}
\description{
Print RDBR model
}
