% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{product_ensemble_votes}
\alias{product_ensemble_votes}
\title{Product vote combination for single-label prediction}
\usage{
product_ensemble_votes(predictions)
}
\arguments{
\item{predictions}{A list of binary predictions.}
}
\value{
A new single-label prediction.
}
\description{
Compute the ensemble prediction using the product votes schema. The
probabilities result is computed using the product of all values.
}
\examples{
model <- br(toyml, "KNN")
predictions <- list(
 predict(model, toyml[1:10], k=1)[, "y1"],
 predict(model, toyml[1:10], k=3)[, "y1"],
 predict(model, toyml[1:10], k=5)[, "y1"]
)

result <- product_ensemble_votes(predictions)
}
\seealso{
Other Ensemble utilites: \code{\link{average_ensemble_votes}},
  \code{\link{majority_ensemble_votes}},
  \code{\link{maximum_ensemble_votes}},
  \code{\link{minimum_ensemble_votes}}
}

