% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_prudent.R
\name{prudent}
\alias{prudent}
\title{PruDent classifier for multi-label Classification}
\usage{
prudent(mdata, base.method = "SVM", phi = 0, ..., save.datasets = FALSE,
  CORES = 1)
}
\arguments{
\item{mdata}{Object of class \code{\link[mldr]{mldr}}, a multi-label train
dataset (provided by \pkg{mldr} package).}

\item{base.method}{A string with the name of base method. The same base method
  will be used for train all subproblems in the two rounds. This is slightly
  different from the original paper that uses two distinct values from this
  parameter.

  Default valid options are: \code{'SVM'}, \code{'C4.5'}, \code{'C5.0'},
  \code{'RF'}, \code{'NB'} and \code{'KNN'}. To use other base method see
  \code{\link{mltrain}} and \code{\link{mlpredict}} instructions. (default:
   \code{'SVM'})}

\item{phi}{A value between 0 and 1 to determine the information gain,
The value 0 include all labels in the second phase and the none.}

\item{...}{Others arguments passed to the base method for all subproblems.}

\item{save.datasets}{Logical indicating whether the binary datasets must be
saved in the model or not. (default: \code{FALSE})}

\item{CORES}{The number of cores to parallelize the training. Values higher
  than 1 require the \pkg{parallel} package. (default: 1)}
}
\value{
An object of class \code{PruDentmodel} containing the set of fitted
  models, including: \describe{
     \item{labels}{A vector with the label names.}
     \item{phi}{The value of \code{phi} parameter.}
     \item{IG}{The matrix of Information Gain used in combination
       with \code{phi} parameter to define the labels used in the second step.
     }
     \item{basemodel}{The BRModel used in the first iteration.}
     \item{metamodels}{A list of models named by the label names used in the
       second iteration.
     }
     \item{datasets}{A list with \code{base} and \code{meta} datasets of
       type \code{mldBR} named by the label names. Only when the
       \code{save.datasets = TRUE}.
     }
 }
}
\description{
Create a PruDent (MBR) classifier to predic multi-label
 data. To this, two round of Binary Relevance is executed, such that,
 the first iteraction generates new attributes to enrich the second
 prediction.

 In the second phase only labels whose information gain is greater than
 a specific phi value is added.
}
\examples{
# Train and predict emotion multilabel dataset using Meta-BR
library(utiml)
testdata <- emotions$dataset[sample(1:100, 10), emotions$attributesIndexes]

# Use SVM as base method
model <- prudent(emotions)
pred <- predict(model, testdata)

# Use different phi correlation with C4.5 classifier
model <- prudent(emotions, 'C4.5', 0.3)
pred <- predict(model, testdata)

# Set a specific parameter
model <- prudent(emotions, 'KNN', k=5)
pred <- predict(model, testdata)
}
\references{
Alali, A., & Kubat, M. (2015). PruDent: A Pruned and Confident Stacking
   Approach for Multi-Label Classification. IEEE Transactions on Knowledge
   and Data Engineering, 27(9), 2480–2493.
}
\seealso{
\code{\link{labels_information_gain}}

Other Transformation methods: \code{\link{brplus}};
  \code{\link{br}}; \code{\link{cc}}; \code{\link{ctrl}};
  \code{\link{dbr}}; \code{\link{ebr}}; \code{\link{ecc}};
  \code{\link{mbr}}; \code{\link{ns}}; \code{\link{rdbr}}
}

