% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threshold.R
\name{rcut_threshold}
\alias{rcut_threshold}
\alias{rcut_threshold.default}
\alias{rcut_threshold.mlresult}
\title{Rank Cut (RCut) threshold method}
\usage{
rcut_threshold(prediction, k)

\method{rcut_threshold}{default}(prediction, k)

\method{rcut_threshold}{mlresult}(prediction, k)
}
\arguments{
\item{prediction}{A matrix or mlresult.}

\item{k}{The number of elements that will be positive.}
}
\value{
A matrix or mlresult based as the type of prediction parameter.
}
\description{
The Rank Cut (RCut) method is an instance-wise strategy, which outputs the k
labels with the highest scores for each instance at the deployment.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Rank Cut (RCut) threshold method for matrix

\item \code{mlresult}: Rank Cut (RCut) threshold method for mlresult
}}
\examples{
prediction <- matrix(runif(16), ncol = 4)
rcut_threshold(prediction, 2)
}
\references{
Al-Otaibi, R., Flach, P., & Kull, M. (2014). Multi-label Classification: A
 Comparative Study on Threshold Selection Methods. In First International
 Workshop on Learning over Multiple Contexts (LMCE) at ECML-PKDD 2014.
}
\seealso{
Other threshold: \code{\link{fixed_threshold}};
  \code{\link{pcut_threshold}},
  \code{\link{pcut_threshold.default}},
  \code{\link{pcut_threshold.mlresult}};
  \code{\link{subset_correction}}
}

