% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_rdbr.R
\name{rdbr}
\alias{rdbr}
\title{Recursive Dependent Binary Relevance (RDBR) for multi-label Classification}
\usage{
rdbr(mdata, base.method = getOption("utiml.base.method", "SVM"),
  estimate.models = TRUE, ..., CORES = getOption("utiml.cores", 1))
}
\arguments{
\item{mdata}{A mldr dataset used to train the binary models.}

\item{base.method}{A string with the name of the base method. (Default:
\code{options("utiml.base.method", "SVM")})}

\item{estimate.models}{Logical value indicatind whether is necessary build
Binary Relevance classifier for estimate process. The default implementaion
use BR as estimators, however when other classifier is desirable then use
the value \code{FALSE} to skip this process. (Default: \code{TRUE}).}

\item{...}{Others arguments passed to the base method for all subproblems.}

\item{CORES}{The number of cores to parallelize the training. Values higher
than 1 require the \pkg{parallel} package. (Default:
\code{options("utiml.cores", 1)})}
}
\value{
An object of class \code{RDBRmodel} containing the set of fitted
 models, including:
 \describe{
   \item{labels}{A vector with the label names.}
   \item{estimation}{The BR model to estimate the values for the labels.
     Only when the \code{estimate.models = TRUE}.}
   \item{models}{A list of final models named by the label names.}
 }
}
\description{
Create a RDBR classifier to predic multi-label data. This is a recursive
approach that enables the binary classifiers to discover existing label
dependency by themselves. The idea of RDBR is running DBR recursivelly until
the results stabilization of the result.
}
\details{
The train method is exactly the same of DBR the recursivity is in the predict
method.
}
\examples{
\dontrun{
# Use SVM as base method
model <- rdbr(toyml)
pred <- predict(model, toyml)

# Use Random Forest as base method and 4 cores
model <- rdbr(toyml, 'RF', CORES = 4)
}
}
\references{
Rauber, T. W., Mello, L. H., Rocha, V. F., Luchi, D., & Varejão, F. M.
  (2014). Recursive Dependent Binary Relevance Model for Multi-label
  Classification. In Advances in Artificial Intelligence - IBERAMIA, 206–217.
}
\seealso{
\code{\link[=dbr]{Dependent Binary Relevance (DBR)}}

Other Transformation methods: \code{\link{brplus}};
  \code{\link{br}}; \code{\link{cc}}; \code{\link{ctrl}};
  \code{\link{dbr}}; \code{\link{ebr}}; \code{\link{ecc}};
  \code{\link{mbr}}; \code{\link{ns}};
  \code{\link{prudent}}
}

