% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{remove_attributes}
\alias{remove_attributes}
\title{Remove attributes from the dataset}
\usage{
remove_attributes(mdata, attributes)
}
\arguments{
\item{mdata}{The mldr dataset to remove labels.}

\item{attributes}{Attributes indexes or attributes names to be removed.}
}
\value{
a new mldr object.
}
\description{
Remove specified attributes generating a new multi-label dataset.
}
\note{
If invalid attributes names or indexes were informed, they will be
 ignored.
}
\examples{
toyml1 <- remove_attributes(toyml, c("iatt8","iatt9", "ratt10"))
toyml2 <- remove_attributes(toyml, 10)
}
\seealso{
Other pre process: 
\code{\link{fill_sparse_mldata}()},
\code{\link{normalize_mldata}()},
\code{\link{remove_labels}()},
\code{\link{remove_skewness_labels}()},
\code{\link{remove_unique_attributes}()},
\code{\link{remove_unlabeled_instances}()},
\code{\link{replace_nominal_attributes}()}
}
\concept{pre process}
