% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{remove_labels}
\alias{remove_labels}
\title{Remove labels from the dataset}
\usage{
remove_labels(mdata, labels)
}
\arguments{
\item{mdata}{The mldr dataset to remove labels.}

\item{labels}{Label indexes or label names to be removed.}
}
\value{
a new mldr object.
}
\description{
Remove spectified labels generating a new multi-label dataset.
}
\examples{
toyml1 <- remove_labels(toyml, c("y1","y5"))
toyml2 <- remove_labels(toyml, c(11, 15))
}
\seealso{
Other pre process: \code{\link{fill_sparce_mldata}},
  \code{\link{normalize_mldata}},
  \code{\link{remove_attributes}},
  \code{\link{remove_skewness_labels}},
  \code{\link{remove_unique_attributes}},
  \code{\link{remove_unlabeled_instances}},
  \code{\link{replace_nominal_attributes}}
}

