% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pre_process.R
\name{remove_skewness_labels}
\alias{remove_skewness_labels}
\title{Remove unusual or very common labels}
\usage{
remove_skewness_labels(mdata, t = 1)
}
\arguments{
\item{mdata}{The mldr dataset to remove the skewness labels.}

\item{t}{Threshold value. Number of minimum examples positive and negative.}
}
\value{
a new mldr object.
}
\description{
Remove the labels that have smaller number of positive or negative examples
based on a specific threshold value.
}
\examples{
remove_skewness_labels(toyml, 20)
}
\seealso{
Other pre process: \code{\link{fill_sparce_mldata}};
  \code{\link{normalize_mldata}};
  \code{\link{remove_attributes}};
  \code{\link{remove_labels}};
  \code{\link{remove_unique_attributes}};
  \code{\link{remove_unlabeled_instances}};
  \code{\link{replace_nominal_attributes}}
}

