% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{remove_unique_attributes}
\alias{remove_unique_attributes}
\title{Remove unique attributes}
\usage{
remove_unique_attributes(mdata)
}
\arguments{
\item{mdata}{The mldr dataset to remove.}
}
\value{
a new mldr object.
}
\description{
Remove the attributes that have a single value for all instances. Empty and
NA values are considered different values.
}
\examples{
alt.toy <- toyml
alt.toy$dataset$ratt10 <- mean(alt.toy$dataset$ratt10)
new.toy <- remove_unique_attributes(alt.toy)
}
\seealso{
Other pre process: 
\code{\link{fill_sparse_mldata}()},
\code{\link{normalize_mldata}()},
\code{\link{remove_attributes}()},
\code{\link{remove_labels}()},
\code{\link{remove_skewness_labels}()},
\code{\link{remove_unlabeled_instances}()},
\code{\link{replace_nominal_attributes}()}
}
\concept{pre process}
