% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{remove_unlabeled_instances}
\alias{remove_unlabeled_instances}
\title{Remove examples without labels}
\usage{
remove_unlabeled_instances(mdata)
}
\arguments{
\item{mdata}{The mldr dataset to remove the instances.}
}
\value{
a new mldr object.
}
\description{
Remove the examples that do not have labels.
}
\examples{
new.toy <- remove_labels(toyml, c(12,14))
remove_unlabeled_instances(new.toy)
}
\seealso{
Other pre process: \code{\link{fill_sparce_mldata}},
  \code{\link{normalize_mldata}},
  \code{\link{remove_attributes}},
  \code{\link{remove_labels}},
  \code{\link{remove_skewness_labels}},
  \code{\link{remove_unique_attributes}},
  \code{\link{replace_nominal_attributes}}
}

