% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{replace_nominal_attributes}
\alias{replace_nominal_attributes}
\title{Replace nominal attributes
Replace the nominal attributes by binary attributes.}
\usage{
replace_nominal_attributes(mdata, ordinal.attributes = list())
}
\arguments{
\item{mdata}{The mldr dataset to remove.}

\item{ordinal.attributes}{Not yet, but it will be used to specify which
attributes need to be replaced.}
}
\value{
a new mldr object.
}
\description{
Replace nominal attributes
Replace the nominal attributes by binary attributes.
}
\examples{
new.toy <- toyml
new.column <- as.factor(sample(c("a","b","c"), 100, replace = TRUE))
new.toy$dataset$ratt10 <- new.column
head(replace_nominal_attributes(new.toy))
}
\seealso{
Other pre process: \code{\link{fill_sparse_mldata}},
  \code{\link{normalize_mldata}},
  \code{\link{remove_attributes}},
  \code{\link{remove_labels}},
  \code{\link{remove_skewness_labels}},
  \code{\link{remove_unique_attributes}},
  \code{\link{remove_unlabeled_instances}}
}
\concept{pre process}
