% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{simple.threshold}
\alias{simple.threshold}
\title{Apply a fixed threshold in the results}
\usage{
simple.threshold(prediction, threshold = 0.5)
}
\arguments{
\item{prediction}{A matrix with scores/probabilities where the columns
are the labels and the rows are the instances.}

\item{threshold}{A single value between 0 and 1 or a list with threshold
   values contained one value per label.}
}
\value{
A matrix with bipartition results
}
\description{
Transfom a prediction matrix with scores/probabilities in
 a bipartion prediction matrix. It is possible use a single value for
 all labels or specify a specific threshold for each label.
}
\examples{
# Create a prediction matrix with scores
result <- matrix(
   data = rnorm(9, 0.5, 0.2),
   ncol = 3,
   dimnames = list(NULL, c("lbl1",  "lb2", "lb3"))
)

# Use 0.5 as threshold
simple.threshold(result)

# Use an threshold for each label
simple.threshold(result, c(0.4, 0.6, 0.7))
}

