% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlresult.R
\name{[.mlresult}
\alias{[.mlresult}
\title{Filter a Multi-Label Result}
\usage{
\method{[}{mlresult}(mlresult, rowFilter = T, colFilter, ...)
}
\arguments{
\item{mlresult}{A mlresult object}

\item{rowFilter}{A list of rows to filter}

\item{colFilter}{A list of columns to filter}

\item{...}{Extra parameters to be used as the filter}
}
\value{
mlresult or matrix. If column filter is performed, then the result
 will be a matrix. Otherwhise, the result will be a mlresult.
}
\description{
If column filter is performed, then the result will be a matrix. Otherwhise,
the result will be a mlresult.
}
