% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{subset.corretion}
\alias{subset.corretion}
\title{Subset Correction of a predicted result}
\usage{
\method{subset}{corretion}(mlresult, train_y, threshold = 0.5)
}
\arguments{
\item{mlresult}{An object of mlresult that contain the scores and
bipartition values.}

\item{train_y}{A matrix/data.frame with all labels values of the training data.}

\item{threshold}{A numeric value between 0 and 1 to use as base to determine
which values needs be reescaled to preserve the corrected labelsets.
(default: 0.5)}
}
\value{
A new mlresult where all results are present in the training labelsets.
}
\description{
This method restrict a multi-label learner prediction to only
 label combinations whose existence is testified by the (training) data. To
 this all labelsets that are predicted but are not found on training data is
 replaced by the most similar labelset.

 If the most simillar is not unique, those label combinations with higher
 frequency in the training data are preferred. The Hamming loss is used
 to determine the difference between the labelsets.
}
\examples{

trainls <- matrix(c(1,1,1, 1,0,1, 1,0,1, 1,0,0, 1,1,1, 1,0,1), ncol = 3, byrow = TRUE)
colnames(trainls) <- c('c1', 'c2', 'c3')

predict <- matrix(c(1,1,0, 0,0,1, 0,0,0), ncol = 3, byrow = TRUE)
colnames(predict) <- c('c1', 'c2', 'c3')

ns.subsetcorrection(predict, trainls)
#       c2 c1 c3
# [1,]  1  1  1
# [2,]  1  0  1
# [3,]  1  0  0
}
\references{
Senge, R., Coz, J. J. del, & Hüllermeier, E. (2013). Rectifying classifier
   chains for multi-label classification. In Workshop of Lernen, Wissen &
   Adaptivität (LWA 2013) (pp. 162–169). Bamberg, Germany.
}
\seealso{
Other threshold: \code{\link{simple.threshold}}
}

