% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_learner.R
\name{summary.mltransformation}
\alias{summary.mltransformation}
\title{Summary method for mltransformation}
\usage{
\method{summary}{mltransformation}(x, ...)
}
\description{
Summary method for mltransformation
}

