% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prediction.R
\name{transform_br_data}
\alias{transform_br_data}
\title{Create a Binary MultiLabel Data}
\usage{
transform_br_data(dataset, classname, base.method, ...)
}
\arguments{
\item{dataset}{A data.frame with the data (the last column must be the class
column)}

\item{classname}{The name of specific class of the object}

\item{base.method}{The name of the base method that will process this dataset}

\item{...}{Extra parameters for adding in the dataset object}
}
\value{
A list with data, labelname, labelindex and methodname.
   This list has three classes: mltransformation, baseMETHODNAME and
   a specific name
}
\description{
Create a Binary MultiLabel Data
}
\examples{
...
tbl <- transform_br_data(dataframe, 'mldBR', 'SVM')
...
}

