% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiml.R
\docType{package}
\name{utiml}
\alias{utiml}
\alias{utiml-package}
\title{utiml: Utilities for Multi-Label Learning}
\description{
The utiml package provides the follow categories of multi-label functions:
\enumerate{
 \item Classification methods
 \item Ensemble utilities
 \item Evaluation measures
 \item Pre-process utilities
 \item Sampling methods
 \item Threshold methods
}
}
\section{Notes}{

 We use the good \code{\link{mldr}} package, to represent and support
 multi-label data. See its documentation to more information about the
 multi-label data.
}
\author{
\itemize{
   \item Adriano Rivolli <rivolli@utfpr.edu.br>
 }
 This package is a result of my PhD at Institute of Mathematics and Computer
 Sciences (ICMC) in the University of Sao Paulo.

 PhD advisor: Andre C. P. L. F. de Carvalho
}

