% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{utiml_all_measures_names}
\alias{utiml_all_measures_names}
\title{MEASURES METHODS ----------------------------------------------------------
Return the tree with the measure names}
\usage{
utiml_all_measures_names()
}
\value{
list
}
\description{
MEASURES METHODS ----------------------------------------------------------
Return the tree with the measure names
}

