% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{utiml_binary_prediction}
\alias{utiml_binary_prediction}
\title{Create a binary.prediction object}
\usage{
utiml_binary_prediction(bipartition, probability)
}
\arguments{
\item{bipartition}{The classes predictions (bipartition values), only 0 and 1}

\item{probability}{The probability/confidence of a prediction, between 0..1}
}
\value{
An object of type binary.prediction
}
\description{
Create a binary.prediction object
}

