% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_compute_binary_ensemble}
\alias{utiml_compute_binary_ensemble}
\title{Compute the ensemble predictions for binary predictions}
\usage{
utiml_compute_binary_ensemble(method.name, binary.predictions)
}
\arguments{
\item{method.name}{The method to compute the votes}

\item{binary.predictions}{A list of binary.prediction}
}
\value{
A new binary.result
}

