% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_compute_ensemble}
\alias{utiml_compute_ensemble}
\title{Compute binary predictions}
\usage{
utiml_compute_ensemble(bipartitions, probabilities, vote.methods, rnames)
}
\arguments{
\item{bipartitions}{A matrix with bipartitions values.}

\item{probabilities}{A matrix with probabilities values.}

\item{vote.methods}{The vote schema method.}

\item{rnames}{The row names.}
}
\value{
A binary.prediction object
}
\description{
Compute binary predictions
}

