% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_compute_ensemble_predictions}
\alias{utiml_compute_ensemble_predictions}
\title{Compute the ensemble predictions based on some vote schema}
\usage{
utiml_compute_ensemble_predictions(predictions, vote.schema)
}
\arguments{
\item{predictions}{A list of matrix predictions}

\item{vote.schema}{Define the way that ensemble must compute the predictions.
The valid options are: \describe{
 \code{'score'}{Compute the averages of probabilities},
 \code{'majority'}{Compute the votes scaled between 0 and \code{m} (number of interations)},
 \code{'prop'}{Compute the proportion of votes, scale data between min and max of votes} }}
}
\value{
A list of mlresult as a result obtained from a multi-label transformation method
}
\description{
Compute the ensemble predictions based on some vote schema
}
\examples{
...
predictions <- list()
predictions$model1 <- prediction(brmodel1, testdata)
predictions$model2 <- prediction(brmodel2, testdata)
result <- utiml_compute_ensemble_predictions(predictions, "majority")
...
}

