% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_compute_multilabel_ensemble}
\alias{utiml_compute_multilabel_ensemble}
\title{Compute the ensemble predictions based on some vote schema}
\usage{
utiml_compute_multilabel_ensemble(predictions, vote.schema,
  probability = TRUE)
}
\arguments{
\item{predictions}{A list of mlresult}

\item{vote.schema}{Define the way that ensemble must compute the predictions.
The valid options are: \describe{
 \code{'MAJ'}{Compute the averages of probabilities},
 \code{'MAX'}{Compute the votes scaled between 0 and \code{m} (number of interations)},
 \code{'MIN'}{Compute the proportion of votes, scale data between min and max of votes}
 \code{'AVG'}{Compute the proportion of votes, scale data between min and max of votes}
 \code{'PROD'}{Compute the proportion of votes, scale data between min and max of votes}
}}

\item{probability}{A logical value. If \code{TRUE} the predicted values are
 the score between 0 and 1, otherwise the values are bipartition 0 or 1.}
}
\value{
A new mlresult that final result
}
\description{
Compute the ensemble predictions based on some vote schema
}
\examples{
...
predictions <- list()
predictions$model1 <- prediction(brmodel1, testdata)
predictions$model2 <- prediction(brmodel2, testdata)
result <- utiml_compute_ensemble_predictions(predictions, "MAJ")
...
}

