% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{utiml_create_binary_data}
\alias{utiml_create_binary_data}
\title{Create a data.frame from original mldr data for a single label}
\usage{
utiml_create_binary_data(mdata, label.name, extra.columns = NULL)
}
\arguments{
\item{mdata}{The mldr dataset}

\item{label.name}{The name of the label}

\item{extra.columns}{New columns to be added in the new dataset}
}
\value{
a data.frame where the label is the last column
}
\description{
Create a data.frame from original mldr data for a single label
}

