% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{utiml_create_model}
\alias{utiml_create_model}
\title{Create Dynamically the model for Binary Relevance Methods}
\usage{
utiml_create_model(dataset, ...)
}
\arguments{
\item{dataset}{An object generated by the \code{\link{utiml_prepare_data}}
method}

\item{...}{Others arguments passed to the base method}
}
\value{
The model dinamically generated
}
\description{
This method is used in the methods to call dinamicaly the base classifier
}

