% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_ensemble_average}
\alias{utiml_ensemble_average}
\title{Average vote combination for a single-label prediction}
\usage{
utiml_ensemble_average(bipartition, probability)
}
\arguments{
\item{bipartition}{A matrix with all bipartition predictions for a single
label. The column are the predictions and the rows the examples.}

\item{probability}{A matrix with all probability predictions for a single
label The column are the predictions and the rows the examples.}
}
\value{
A list with two values "bipartition" and "probability".
}
\description{
Compute the prediction for a single-label using the average votes schema.
The probabilities result is computed using the averaged values.
}
