% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_ensemble_average_votes}
\alias{utiml_ensemble_average_votes}
\title{Average vote combination for ensemble prediction}
\usage{
utiml_ensemble_average_votes(predictions)
}
\arguments{
\item{predictions}{A list of multi-label predictions
 (a \code{\link{mlprediction}} object).}
}
\value{
A multi-label prediction (a \code{\link{mlprediction}} object).
}
\description{
Compute the ensemble prediction using the average
 votes schema. The probabilities result is computed using the
 averaged values.
}
\examples{
predictions <- list(
 as.binaryPrediction(c(1,1,1,1)),
 as.binaryPrediction(c(0.6,0.1,0.7,0.2)),
 as.binaryPrediction(c(0.8,0.1,0.4,0.3))
)
result <- utiml_ensemble_average_votes(predictions)
result$probability
# --> 0.8, 0.4, 0.7, 0.5
}
\seealso{
\code{link{mlprediction}}

Other Ensemble utilites: \code{\link{utiml_ensemble_majority_votes}};
  \code{\link{utiml_ensemble_maximum_votes}};
  \code{\link{utiml_ensemble_minimum_votes}};
  \code{\link{utiml_ensemble_product_votes}}
}

