% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_ensemble_compute_votes}
\alias{utiml_ensemble_compute_votes}
\title{Generic vote combination for ensemble prediction}
\usage{
utiml_ensemble_compute_votes(predictions, method.name)
}
\arguments{
\item{predictions}{A list of multi-label predictions
(a \code{\link{mlprediction}} object).}

\item{method.name}{The name or the method that will be compute
 the results. This method receive a list of probability values
 and must return a single value probability.}
}
\value{
A multi-label prediction (a \code{\link{mlprediction}} object).
}
\description{
Generic vote combination for ensemble prediction
}
\examples{
predictions <- list(
 as.binaryPrediction(c(1,1,1,1)),
 as.binaryPrediction(c(0.6,0.1,0.7,0.2)),
 as.binaryPrediction(c(0.8,0.1,0.4,0.3))
)

# Compute the maximum vote combination
result <- utiml_ensemble_compute_votes(predictions, max)
}
\seealso{
\code{link{mlprediction}}
}

