% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_ensemble_majority_votes}
\alias{utiml_ensemble_majority_votes}
\title{Majority vote combination for ensemble prediction}
\usage{
utiml_ensemble_majority_votes(predictions)
}
\arguments{
\item{predictions}{A list of multi-label predictions
(a \code{\link{mlprediction}} object).}
}
\value{
A multi-label prediction (a \code{\link{mlprediction}} object).
}
\description{
Compute the ensemble prediction using the majority
 votes schema. The probabilities result is computed using only
 the majority instances. In others words, if a example is
 predicted as posivite, only the positive confidences are used to
 compute the averaged value.
}
\examples{
predictions <- list(
 as.binaryPrediction(c(1,1,1,1)),
 as.binaryPrediction(c(0.6,0.1,0.8,0.2)),
 as.binaryPrediction(c(0.8,0.3,0.4,0.1))
)
result <- utiml_ensemble_majority_votes(predictions)
}
\seealso{
\code{link{mlprediction}}

Other Ensemble utilites: \code{\link{utiml_ensemble_average_votes}},
  \code{\link{utiml_ensemble_maximum_votes}},
  \code{\link{utiml_ensemble_minimum_votes}},
  \code{\link{utiml_ensemble_product_votes}}
}

