% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_ensemble_method}
\alias{utiml_ensemble_method}
\title{Define the method name related with the vote schema}
\usage{
utiml_ensemble_method(vote.schema)
}
\arguments{
\item{vote.schema}{Define the way that ensemble must compute the predictions.}
}
\value{
The method name that will compute the votes
}
\description{
Define the method name related with the vote schema
}
