% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_ensemble_minimum}
\alias{utiml_ensemble_minimum}
\title{Minimum vote combination for single-label prediction}
\usage{
utiml_ensemble_minimum(bipartition, probability)
}
\arguments{
\item{bipartition}{A matrix with all bipartition predictions for a single
label. The column are the predictions and the rows the examples.}

\item{probability}{A matrix with all probability predictions for a single
label The column are the predictions and the rows the examples.}
}
\value{
A list with two values "bipartition" and "probability".
}
\description{
Compute the single-label prediction using the minimum votes schema. The
probabilities result is computed using the minimum value.
}
