% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_ensemble_product_votes}
\alias{utiml_ensemble_product_votes}
\title{Product vote combination for ensemble prediction}
\usage{
utiml_ensemble_product_votes(predictions)
}
\arguments{
\item{predictions}{A list of multi-label predictions
(a \code{\link{mlprediction}} object).}
}
\value{
A multi-label prediction (a \code{\link{mlprediction}} object).
}
\description{
Compute the ensemble prediction using the product
 votes schema. The probabilities result is computed using the
 product of all values.
}
\examples{
predictions <- list(
 as.binaryPrediction(c(1,1,0.5,1)),
 as.binaryPrediction(c(0.9,0.1,0.5,0.2)),
 as.binaryPrediction(c(0.8,0.5,0.5,0.3))
)
result <- utiml_ensemble_product_votes(predictions)
result$probability
# --> 0.72, 0.05, 0.125, 0.06
}
\seealso{
\code{link{mlprediction}}

Other Ensemble utilites: \code{\link{utiml_ensemble_average_votes}},
  \code{\link{utiml_ensemble_majority_votes}},
  \code{\link{utiml_ensemble_maximum_votes}},
  \code{\link{utiml_ensemble_minimum_votes}}
}

