% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_get_schema_method}
\alias{utiml_get_schema_method}
\title{Define the method related with the vote schema}
\usage{
utiml_get_schema_method(vote.schema)
}
\arguments{
\item{vote.schema}{Define the way that ensemble must compute the predictions.}
}
\value{
The method that will compute the votes
}
\description{
Define the method related with the vote schema
}

