% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_ifelse}
\alias{utiml_ifelse}
\title{Conditional value selection}
\usage{
utiml_ifelse(test, yes, no)
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}

\item{yes}{return values for true elements of test.}

\item{no}{return values for false elements of test.}
}
\value{
The respective value yes or no based on test. This
 is an alternative to use a single logical value for
 lists, matrices and other composed data
}
\description{
Conditional value selection
}
\examples{
utiml_ifelse(TRUE, dataframe1, dataframe2) #dataframe1
utiml_ifelse(length(my.list) > 10, my.list[1:10], my.list)
}

