% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_is_equal_sets}
\alias{utiml_is_equal_sets}
\title{Define if two sets are equals independently of the order of the elements}
\usage{
utiml_is_equal_sets(a, b)
}
\arguments{
\item{a}{A list}

\item{b}{Other list}
}
\value{
Logical value where TRUE the sets are equals and FALSE otherwise.
}
\description{
Define if two sets are equals independently of the order of the elements
}
\examples{
utiml_is_equal_sets(c(1, 2, 3), c(3, 2, 1))
## TRUE

utiml_is_equal_sets(c(1, 2, 3), c(1, 2, 3, 4))
## FALSE
}

