% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{utiml_kfold}
\alias{utiml_kfold}
\title{Create the k partitions of k-fold}
\usage{
utiml_kfold(mdata, k, kfold.method)
}
\arguments{
\item{mdata}{A dataset of class \code{\link[mldr]{mldr}}.}

\item{k}{The number of folds.}

\item{kfold.method}{The method to split the data.}
}
\value{
An object of type mldr_kfolds
}
\description{
Create the k partitions of k-fold
}
\examples{
utiml_kfold(mdata, 10, utiml_random_split)
}
\seealso{
Other sampling: \code{\link{mldr_getfold}};
  \code{\link{mldr_iterative_stratification_holdout}};
  \code{\link{mldr_iterative_stratification_kfold}};
  \code{\link{mldr_random_holdout}};
  \code{\link{mldr_random_kfold}};
  \code{\link{mldr_stratified_holdout}};
  \code{\link{mldr_stratified_kfold}};
  \code{\link{utiml_holdout}}
}

