% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_prudent.R
\name{utiml_labels_IG}
\alias{utiml_labels_IG}
\title{Calculate the Information Gain for each pair of labels}
\usage{
utiml_labels_IG(mdata)
}
\arguments{
\item{mdata}{A mldr dataset containing the label information.}
}
\value{
A matrix where the rows and columns represents the labels.
}
\description{
Calculate the Information Gain for each pair of labels
}
\references{
Alali, A., & Kubat, M. (2015). PruDent: A Pruned and Confident Stacking
  Approach for Multi-Label Classification. IEEE Transactions on Knowledge
  and Data Engineering, 27(9), 2480-2493.
}

