% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_labelset_stratification}
\alias{utiml_labelset_stratification}
\title{Labelsets Stratification}
\usage{
utiml_labelset_stratification(mdata, r)
}
\arguments{
\item{mdata}{A mldr dataset}

\item{r}{Desired proportion of examples in each subset, r1, . . . rk}
}
\value{
A list with k disjoint indexes subsets S1, . . .Sk
}
\description{
Create the indexes using the Labelsets Stratification
  approach.
}
\examples{
# Create 3 partitions for train, validation and test
indexes <- utiml_labelset_stratification(emotions, c(0.6,0.1,0.3))

# Create a stratified 10-fold
indexes <- utiml_labelset_stratification(emotions, rep(0.1,10))
}

