% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_lapply}
\alias{utiml_lapply}
\title{Select the suitable method lapply or mclaplly}
\usage{
utiml_lapply(mylist, myfnc, utiml.cores, utiml.seed = NA, ...)
}
\arguments{
\item{mylist}{a list to iterate.}

\item{myfnc}{The function to be applied to each element of the mylist.}

\item{utiml.cores}{The number of cores to use. If 1 use lapply oterwise use
mclapply.}

\item{utiml.seed}{A numeric value to set a seed to execute in parallel mode.}

\item{...}{Extra arguments to myfnc.}
}
\value{
A list with the results of the specified method.
}
\description{
Select the suitable method lapply or mclaplly
}

