% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{utiml_measure_binary_balacc}
\alias{utiml_measure_binary_balacc}
\title{Compute the binary balanced accuracy}
\usage{
utiml_measure_binary_balacc(TP, FP, TN, FN)
}
\arguments{
\item{TP}{The number of True Positive values}

\item{FP}{The number of False Positive values}

\item{TN}{The number of True Negative values}

\item{FN}{The number of False Negative values}
}
\value{
Balanced accuracy value between 0 and 1
}
\description{
Compute the binary balanced accuracy
}
