% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{utiml_measure_coverage}
\alias{utiml_measure_coverage}
\title{Multi-label Coverage Measure}
\usage{
utiml_measure_coverage(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}

\item{...}{ignored}
}
\description{
Multi-label Coverage Measure
}
\references{
Schapire, R. E., & Singer, Y. (2000). BoosTexter: A boosting-
based system for text categorization. Machine Learning, 39(2), 135-168.
}
