% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{utiml_measure_is_error}
\alias{utiml_measure_is_error}
\title{Multi-label Is Error Measure}
\usage{
utiml_measure_is_error(mlconfmat, ranking, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}

\item{ranking}{The expected matrix ranking}

\item{...}{ignored}
}
\description{
Multi-label Is Error Measure
}
\references{
Crammer, K., & Singer, Y. (2003). A Family of Additive Online
 Algorithms for Category Ranking. Journal of Machine Learning Research, 3(6),
 1025-1058.
}

