% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{utiml_measure_micro_AUC}
\alias{utiml_measure_micro_AUC}
\title{Multi-label Macro-AUC Measure}
\usage{
utiml_measure_micro_AUC(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}

\item{...}{ignored}
}
\description{
Multi-label Macro-AUC Measure
}
\references{
Zhang, M.-L., & Zhou, Z.-H. (2014). A Review on Multi-Label
 Learning Algorithms. IEEE Transactions on Knowledge and Data Engineering,
 26(8), 1819-1837.
}

