% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{utiml_measure_micro_recall}
\alias{utiml_measure_micro_recall}
\title{Multi-label Micro-Recall Measure}
\usage{
utiml_measure_micro_recall(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}

\item{...}{ignored}
}
\description{
Multi-label Micro-Recall Measure
}
\references{
Gibaja, E., & Ventura, S. (2015). A Tutorial on Multilabel
 Learning. ACM Comput. Surv., 47(3), 52:1-52:38.
}
