% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{utiml_measure_names}
\alias{utiml_measure_names}
\title{Return the name of measures}
\usage{
utiml_measure_names(measures = c("all"))
}
\arguments{
\item{measures}{The group of measures (Default: "all").}
}
\value{
array of character contained the measures names.
}
\description{
Return the name of measures
}
\examples{
\dontrun{
utiml_measure_names()
utiml_measure_names("bipartition")
utiml_measure_names(c("micro-based", "macro-based"))
}
}
