% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{utiml_measure_ranking_error}
\alias{utiml_measure_ranking_error}
\title{Multi-label Ranking Error Measure}
\usage{
utiml_measure_ranking_error(mlconfmat, ranking, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}

\item{ranking}{A matrix ranking}

\item{...}{ignored}
}
\description{
Multi-label Ranking Error Measure
}
\references{
Park, S.-H., & Furnkranz, J. (2008). Multi-Label Classification
 with Label Constraints. Proceedings of the ECML PKDD 2008 Workshop on
 Preference Learning (PL-08, Antwerp, Belgium), 157-171.
}
