% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_newdata}
\alias{utiml_newdata}
\title{Return the newdata to a data.frame or matrix}
\usage{
utiml_newdata(newdata)
}
\arguments{
\item{newdata}{The data.frame or mldr data}
}
\value{
A dataframe or matrix containing only dataset
}
\examples{
test <- emotions$dataset[,emotions$attributesIndexes]
all(test == utiml_newdata(emotions)) # TRUE
all(test == utiml_newdata(test)) # TRUE
}

