% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_normalize}
\alias{utiml_normalize}
\title{Internal normalize data function}
\usage{
utiml_normalize(data, max.val = NULL, min.val = NULL)
}
\arguments{
\item{data}{a set of numbers.}

\item{max.val}{The maximum value to normalize. If NULL use the max value
present in the data. (default: \code{NULL})}

\item{min.val}{The minimum value to normalize. If NULL use the min value
present in the data (default: \code{NULL})}
}
\value{
The normalized data
}
\description{
Internal normalize data function
}
\examples{
\dontrun{
utiml_normalize(c(1,2,3,4,5))
#--> 0 0.25 0.5 0.75 1

utiml_normalize(c(1,2,3,4,5), 10, 0)
#--> 0.1 0.2 0.3 0.4 0.5
}
}

