% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{utiml_predict}
\alias{utiml_predict}
\title{Create a predictive multi-label result}
\usage{
utiml_predict(predictions, probability)
}
\arguments{
\item{predictions}{The list of binary.prediction obtained from union of all
binary predictions.}

\item{probability}{A logical value. If \code{TRUE} the predicted values are
the score between 0 and 1, otherwise the values are bipartition 0 or 1.}
}
\value{
A matrix containing the probabilistic values or just predictions.
 If the matrix contains the probabilistic values then an attribute called
 'classes' contains the bipartitions values. Otherwise, if the matrix
 contains the bipartitions values then an attribute called 'probs' cantains
 the probabilities
}
\description{
This function select the correct result and organize them in a prediction
 matrix where the columns are the labels and the rows are the test examples.
 If probability is \code{TRUE} the values contain the labels probabilities,
 otherwise the values are the predictive value '0' or '1'.
}

