% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_predict_binary_ensemble}
\alias{utiml_predict_binary_ensemble}
\title{Predict binary predictions}
\usage{
utiml_predict_binary_ensemble(predictions, vote.schema)
}
\arguments{
\item{predictions}{A list of binary predictions.}

\item{vote.schema}{The name of vote schema.}
}
\value{
A binary.prediction object
}
\description{
Is very similar from utiml_compute_ensemble but differs from arguments
}
