% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{utiml_predict_binary_model}
\alias{utiml_predict_binary_model}
\title{Dinamically call the prediction function}
\usage{
utiml_predict_binary_model(model, newdata, ...)
}
\arguments{
\item{model}{A model generated by \code{\link{utiml_create_model}} method}

\item{newdata}{The data to be predicted}

\item{...}{Others arguments passed to the base method}
}
\value{
A binary.prediction object
}
\description{
Dinamically call the prediction function
}

