% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampling.R
\name{utiml_random_split}
\alias{utiml_random_split}
\title{Random split of a dataset}
\usage{
utiml_random_split(mdata, r)
}
\arguments{
\item{mdata}{A mldr dataset.}

\item{r}{Desired proportion of examples in each subset r1, . . . rk.}
}
\value{
A list with k disjoint indexes subsets S1, . . .Sk.
}
\description{
Random split of a dataset
}
\examples{
utiml_random_split(emotions, c(0.6, 0.2, 0.2))
}

