% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_rename}
\alias{utiml_rename}
\title{Rename the list using the names values or its own content}
\usage{
utiml_rename(X, names = NULL)
}
\arguments{
\item{X}{A list}

\item{names}{The list names, If empty the content of X is used}
}
\value{
A list with the new names
}
\description{
Rename the list using the names values or its own content
}
\examples{
utiml_rename(c("a", "b", "c"))
## c(a="a", b="b", c="c")

utiml_rename(c(1, 2, 3), c("a", "b", "c"))
## c(a=1, b=2, c=3)
}

