% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{utiml_validate_splitmethod}
\alias{utiml_validate_splitmethod}
\title{Return the name of split method and validate if it is valid}
\usage{
utiml_validate_splitmethod(method)
}
\arguments{
\item{method}{The method name}
}
\value{
The correct name of split method
}
\description{
Return the name of split method and validate if it is valid
}
