% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_vote.schema_method}
\alias{utiml_vote.schema_method}
\title{Define the method related with the vote schema}
\usage{
utiml_vote.schema_method(vote.schema)
}
\arguments{
\item{vote.schema}{Define the way that ensemble must compute the predictions.
 The valid options are:
\describe{
 \code{'MAJ'}{Compute the averages of probabilities},
 \code{'MAX'}{Compute the votes scaled between 0 and \code{m} (number of interations)},
 \code{'MIN'}{Compute the proportion of votes, scale data between min and max of votes}
 \code{'AVG'}{Compute the proportion of votes, scale data between min and max of votes}
 \code{'PROD'}{Compute the proportion of votes, scale data between min and max of votes}
}}
}
\value{
The method that will compute the votes
}
\description{
Define the method related with the vote schema
}

