% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{as.binaryPrediction}
\alias{as.binaryPrediction}
\title{Create a predictive binary result object}
\usage{
as.binaryPrediction(probability, threshold = 0.5)
}
\arguments{
\item{probability}{A vector with probabilities predictions or with bipartitions
prediction for a binary prediction.}

\item{threshold}{A numeric value between 0 and 1 to create the bipartitions.}
}
\value{
An object of type "\code{binary.prediction}" used by problem transformation
 methods that use binary classifiers. It has only two attributes:
 \code{bipartition} and \code{probability}, that respectively have the
 bipartition and probabilities results.
}
\description{
The binary transformation methods require a specific data format from
 base classifiers prediction this is used to transform the result from the base
 methods.
}
\examples{
probs <- runif(nrow(newdata), 0, 1)
as.binaryPrediction(probs)

# Use different threshold value
probs <- runif(10, 0, 1)
result <- as.binaryPrediction(probs, 0.6)
}

