% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{as.bipartition}
\alias{as.bipartition}
\title{Convert a mlresult to a matrix with bipartition values}
\usage{
as.bipartition(mlresult)
}
\arguments{
\item{mlresult}{The mlresult object}
}
\value{
matrix with bipartition values
}
\description{
Convert a mlresult to a matrix with bipartition values
}

