% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{as.mlresult}
\alias{as.mlresult}
\alias{as.mlresult.default}
\alias{as.mlresult.matrix}
\alias{as.mlresult.mlresult}
\title{Convert a matrix prediction in a multi label prediction}
\usage{
as.mlresult(predictions, ...)

\method{as.mlresult}{default}(predictions, probability = T, threshold = 0.5,
  ...)

\method{as.mlresult}{matrix}(predictions, probability = T, threshold = 0.5,
  ...)

\method{as.mlresult}{mlresult}(predictions, probability = T, ...)
}
\arguments{
\item{predictions}{a Matrix or data.frame contained the scores/probabilities
values}

\item{...}{Others parameters passed to the method as threshold}

\item{probability}{A logical value. If \code{TRUE} the predicted values are
the score between 0 and 1, otherwise the values are bipartition 0 or 1.}

\item{threshold}{Threshold value for create bipartition (Default: 0.5)}
}
\value{
An object of type mlresult
}
\description{
Convert a matrix prediction in a multi label prediction
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default mlresult transform method

\item \code{matrix}: Matrix mlresult transform method

\item \code{mlresult}: change the mlresult type
}}
\examples{
predictions <- matrix(sample(1:1000, 100, replace = TRUE) / 1000, ncol = 10)
colnames(predictions) <- paste('label', 1:10, sep='')

as.mlresult(predictions)
as.mlresult(predictions, probability = FALSE)
as.mlresult(predictions, threshold = 0.6)
}

