% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlresult.R
\name{as.probability}
\alias{as.probability}
\title{Convert a mlresult to a probability matrix}
\usage{
as.probability(mlresult)
}
\arguments{
\item{mlresult}{The mlresult object}
}
\value{
matrix with probabilities values
}
\description{
Convert a mlresult to a probability matrix
}

