% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{as.resultPrediction}
\alias{as.resultPrediction}
\title{Create a predictive result object}
\usage{
as.resultPrediction(probability, threshold = 0.5)
}
\arguments{
\item{probability}{A vector with probabilities predictions or with bipartitions
prediction for a binary prediction.}

\item{threshold}{A numeric value betweenpredictions[["class1"]] <- mlpredict(model1, testdata) 0 and 1 to create the bipartitions.}
}
\value{
An object of type "\code{mlresult}" used by problem transformation
 methods that use binary classifiers. It has only two attributes:
 \code{bipartition} and \code{probability}, that respectively have the
 bipartition and probabilities results.
}
\description{
The transformation methods require a specific data format from base
 classifiers prediDefine the way that ensemble must compute the predictions.
The valid options are: \describe{
 \code{'score'}{Compute the averages of probabilities},
 \code{'majority'}{Compute the votes scaled between 0 and \code{m} (number of interations)},
 \code{'prop'}{Compute the proportion of votes, scale data between min and max of votes} }ction. If you implement a new base method then use this method
 to return the final result of your \code{mlpredict} method.
}
\examples{
# This method is used to implement a mlpredict based method
# In this example we create a random predict method
mlpredicti.random <- function (model, newdata, ...) {
   probs <- runif(nrow(newdata), 0, 1)
   as.resultPrediction(probs)
}

# Define a different threshold for a specific subproblem use
...Define the way that ensemble must compute the predictions.
The valid options are: \\describe{
 \\code{'score'}{Compute the averages of probabilities},
 \\code{'majority'}{Compute the votes scaled between 0 and \\code{m} (number of interations)},
 \\code{'prop'}{Compute the proportion of votes, scale data between min and max of votes} }
result <- as.resultPrediction(probs, 0.6)
...
}

