% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{br.create_model}
\alias{br.create_model}
\title{Create Dynamically the model for Binary Relevance Methods}
\usage{
br.create_model(dataset, ...)
}
\arguments{
\item{dataset}{An object generated by the \code{\link{br.transformation}} method}

\item{...}{Others arguments passed to the base method}
}
\value{
The model dinamically generated
}
\description{
Create Dynamically the model for Binary Relevance Methods
}
\examples{
# Create a model for each dataset in datasets list
lapply(datasets, br.create_model, ...)
}

