% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br.R
\name{br}
\alias{br}
\title{Binary Relevance for multi-label Classification}
\usage{
br(mdata, base.method = getOption("utiml.cores", "SVM"), ...,
  CORES = getOption("utiml.cores", 1))
}
\arguments{
\item{mdata}{A mldr dataset used to train the binary models.}

\item{base.method}{A string with the name of the base method. (Default:
\code{options("utiml.base.method", "SVM")})}

\item{...}{Others arguments passed to the base method for all subproblems}

\item{CORES}{The number of cores to parallelize the training. Values higher
than 1 require the \pkg{parallel} package. (Default:
\code{options("utiml.cores", 1)})}
}
\value{
An object of class \code{BRmodel} containing the set of fitted
  models, including:
  \describe{
   \item{labels}{A vector with the label names}
   \item{models}{A list of the generated models, named by the label names.}
  }
}
\description{
Create a Binary Relevance model for multilabel classification.
}
\details{
Binary Relevance is a simple and effective transformation method to predict
multi-label data. This is based on the one-versus-all approach to build a
specific model for each label.
}
\examples{
\dontrun {
# Use SVM as base method
model <- br(toyml)
pred <- predict(model, toyml)

# Change the default base method and use 4 CORES
model <- br(toyml[1:50], 'RF', CORES = 4)

# Set a parameters for all subproblems
model <- br(toyml, 'KNN', k=5)
}
}
\references{
Boutell, M. R., Luo, J., Shen, X., & Brown, C. M. (2004). Learning
   multi-label scene classification. Pattern Recognition, 37(9), 1757–1771.
}
\seealso{
Other Transformation methods: \code{\link{brplus}},
  \code{\link{cc}}, \code{\link{ctrl}}, \code{\link{dbr}},
  \code{\link{ebr}}, \code{\link{ecc}}, \code{\link{mbr}},
  \code{\link{ns}}, \code{\link{prudent}},
  \code{\link{rdbr}}
}

