% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/brplus.R
\name{brplus}
\alias{brplus}
\title{BR+ or BRplus for multi-label Classification}
\usage{
brplus(mdata, base.method = "SVM", ..., save.datasets = FALSE, CORES = 1)
}
\arguments{
\item{mdata}{Object of class \code{\link[mldr]{mldr}}, a multi-label train
dataset (provided by \pkg{mldr} package).}

\item{base.method}{A string with the name of base method. The same base method
  will be used for train all subproblems and the BR classifers

  Default valid options are: \code{'SVM'}, \code{'C4.5'}, \code{'C5.0'},
  \code{'RF'}, \code{'NB'} and \code{'KNN'}. To use other base method see
  \code{\link{mltrain}} and \code{\link{mlpredict}} instructions. (default:
   \code{'SVM'}).}

\item{...}{Others arguments passed to the base method for all subproblems.}

\item{save.datasets}{Logical indicating whether the binary datasets must be
saved in the model or not. (default: FALSE)}

\item{CORES}{he number of cores to parallelize the training. Values higher
  than 1 require the \pkg{parallel} package. (default: 1)}
}
\value{
An object of class \code{BRPmodel} containing the set of fitted
  models, including: \describe{
   \item{freq}{The label frequencies to use with the "Stat" strategy}
   \item{initial}{The BR model to predict the values for the labels to initial step}
   \item{models}{A list of final models named by the label names.}
   \item{datasets}{A list with \code{initial} and \code{final} datasets of
     type \code{mldBRP} named by the label names. Only when the
     \code{save.datasets = TRUE}.
   }
}
}
\description{
Create a BR+ classifier to predic multi-label data.
 This is a simple approach that enables the binary classifiers to
 discover existing label dependency by themselves. The main idea of
 BR+ is to increment the feature space of the binary classifiers to
 let them discover exist- ing label dependency by themselves.

 This implementation has different strategy to predict the final
 set of labels for unlabeled examples, as proposed in original
 paper.
}
\examples{
# Train and predict emotion multilabel dataset using BRPlus
library(utiml)
testdata <- emotions$dataset[sample(1:100, 10), emotions$attributesIndexes]

# Use SVM as base method
model <- brplus(emotions)
pred <- predict(model, testdata)

# Use Random Forest as base method and 4 cores
model <- brplus(emotions, "RF", CORES = 4)
pred <- predict(model, testdata)
}
\references{
Cherman, E. A., Metz, J., & Monard, M. C. (2012). Incorporating label
   dependency into the binary relevance framework for multi-label
   classification. Expert Systems with Applications, 39(2), 1647–1655.
}
\seealso{
Other Transformation methods: \code{\link{br}};
  \code{\link{cc}}; \code{\link{dbr}}; \code{\link{ebr}};
  \code{\link{ecc}}; \code{\link{mbr}}; \code{\link{ns}}
}

