% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Classifier Chains for multi-label Classification}
\usage{
cc(mdata, base.method = "SVM", chain = c(), ..., CORES = 1)
}
\arguments{
\item{mdata}{Object of class \code{\link[mldr]{mldr}}, a multi-label train
dataset (provided by \pkg{mldr} package).}

\item{base.method}{A string with the name of base method. The same base method
  will be used for train all subproblems.

  Default valid options are: \code{'SVM'}, \code{'C4.5'}, \code{'C5.0'},
  \code{'RF'}, \code{'NB'} and \code{'KNN'}. To use other base method see
  \code{\link{mltrain}} and \code{\link{mlpredict}} instructions. (default:
   \code{'SVM'})}

\item{chain}{A vector with the label names to define the chain order. If
empty the chain is the default label sequence of the dataset. (default:
\code{list()})}

\item{...}{Others arguments passed to the base method for all subproblems.}

\item{CORES}{he number of cores to parallelize the training. Values higher
  than 1 require the \pkg{parallel} package. (default: 1)}
}
\value{
An object of class \code{CCmodel} containing the set of fitted
  models, including: \describe{
  \item{chain}{A vector with the chain order}
  \item{labels}{A vector with the label names in expected order}
  \item{models}{A list of models named by the label names.}
}
}
\description{
Create a Classifier Chains model for multilabel classification.

  Classifier Chains is a Binary Relevance transformation method based to predict
  multi-label data. This is based on the one-versus-all approach to build a
  specific model for each label. It is different from BR method due the strategy
  of extended the attribute space with the 0/1 label relevances of all previous
  classifiers, forming a classifier chain.
}
\examples{
# Train and predict emotion multilabel dataset using Classifier Chains
dataset <- mldr_random_holdout(emotions, c(train=0.9, test=0.1))

# Use SVM as base method
model <- cc(dataset$train)
pred <- predict(model, dataset$test)

# Use a specific chain with C4.5 classifier
mychain <- sample(rownames(dataset$train$labels))
model <- cc(dataset$train, 'C4.5', mychain)
pred <- predict(model, dataset$test)

# Set a specific parameter
model <- cc(dataset$train, 'KNN', k=5)
pred <- predict(model, dataset$test)
}
\references{
Read, J., Pfahringer, B., Holmes, G., & Frank, E. (2011). Classifier chains
   for multi-label classification. Machine Learning, 85(3), 333–359.

 Read, J., Pfahringer, B., Holmes, G., & Frank, E. (2009). Classifier Chains
   for Multi-label Classification. Machine Learning and Knowledge Discovery
   in Databases, Lecture Notes in Computer Science, 5782, 254–269.
}
\seealso{
Other Transformation methods: \code{\link{brplus}};
  \code{\link{br}}; \code{\link{ctrl}}; \code{\link{dbr}};
  \code{\link{ebr}}; \code{\link{ecc}}; \code{\link{mbr}};
  \code{\link{ns}}; \code{\link{prudent}};
  \code{\link{rdbr}}
}

