% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Classifier Chains for multi-label Classification}
\usage{
cc(mdata, base.method = "SVM", chain = c(), ..., save.datasets = FALSE,
  CORES = 1)
}
\arguments{
\item{mdata}{Object of class \code{\link[mldr]{mldr}}, a multi-label train
dataset (provided by \pkg{mldr} package).}

\item{base.method}{A string with the name of base method. The same base method
  will be used for train all subproblems.

  Default valid options are: \code{'SVM'}, \code{'C4.5'}, \code{'C5.0'},
  \code{'RF'}, \code{'NB'} and \code{'KNN'}. To use other base method see
  \code{\link{mltrain}} and \code{\link{mlpredict}} instructions. (default:
   \code{'SVM'})}

\item{chain}{A vector with the label names to define the chain order. If
empty the chain is the default label sequence of the dataset. (default:
\code{list()})}

\item{...}{Others arguments passed to the base method for all subproblems.}

\item{save.datasets}{Logical indicating whether the binary datasets must be
saved in the model or not. (default: FALSE)}

\item{CORES}{he number of cores to parallelize the training. Values higher
  than 1 require the \pkg{parallel} package. (default: 1)}
}
\value{
An object of class \code{CCmodel} containing the set of fitted
  models, including: \describe{ \item{chain}{A vector with the chain order}
  \item{labels}{A vector with the label names in expected order}
  \item{models}{A list of models named by the label names.} \item{datasets}{A
  list of \code{mldCC} named by the label names. Only when the
  \code{save.datasets = TRUE}.} }
}
\description{
Create a Classifier Chains model for multilabel classification.

  Classifier Chains is a Binary Relevance transformation method based to predict
  multi-label data. This is based on the one-versus-all approach to build a
  specific model for each label. It is different from BR method due the strategy
  of extended the attribute space with the 0/1 label relevances of all previous
  classifiers, forming a classifier chain.
}
\section{Warning}{

   RWeka package does not permit use \code{'C4.5'} in parallel mode, use
   \code{'C5.0'} or \code{'CART'} instead of it.
}
\examples{
# Train and predict emotion multilabel dataset using Classifier Chains
library(utiml)
testdata <- emotions$dataset[sample(1:100, 10), emotions$attributesIndexes]

# Use SVM as base method
model <- cc(emotions)
pred <- predict(model, testdata)

# Use a specific chain with C4.5 classifier
mychain <- sample(rownames(emotions$labels))
model <- cc(emotions, "C4.5", mychain)
pred <- predict(model, testdata)

# Set a specific parameter
model <- cc(emotions, "KNN", k=5)
pred <- predict(model, testdata)
}
\references{
Read, J., Pfahringer, B., Holmes, G., & Frank, E. (2011). Classifier chains
   for multi-label classification. Machine Learning, 85(3), 333–359.
}
\seealso{
Other Transformation methods: \code{\link{brplus}};
  \code{\link{br}}; \code{\link{dbr}}; \code{\link{ebr}};
  \code{\link{ecc}}; \code{\link{mbr}}; \code{\link{ns}}
}

