% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensemble.R
\name{check_ensemble_vote}
\alias{check_ensemble_vote}
\title{Verify if a schema vote name is valid}
\usage{
check_ensemble_vote(vote.schema, accept.null = TRUE)
}
\description{
Verify if a schema vote name is valid
}

