% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_ebr.R
\name{ebr}
\alias{ebr}
\title{Ensemble of Binary Relevance for multi-label Classification}
\usage{
ebr(mdata, base.method = getOption("utiml.base.method", "SVM"), m = 10,
  subsample = 0.75, attr.space = 0.5, replacement = TRUE, ...,
  cores = getOption("utiml.cores", 1), seed = getOption("utiml.seed", NA))
}
\arguments{
\item{mdata}{A mldr dataset used to train the binary models.}

\item{base.method}{A string with the name of the base method. (Default:
\code{options("utiml.base.method", "SVM")})}

\item{m}{The number of Binary Relevance models used in the ensemble.}

\item{subsample}{A value between 0.1 and 1 to determine the percentage of
training instances that must be used for each classifier. (Default: 0.75)}

\item{attr.space}{A value between 0.1 and 1 to determine the percentage of
attributes that must be used for each classifier. (Default: 0.50)}

\item{replacement}{Bollean value to define if use sampling with replacement
to create the data of the models of the ensemble. (Default: TRUE)}

\item{...}{Others arguments passed to the base method for all subproblems}

\item{cores}{The number of cores to parallelize the training. Values higher
than 1 require the \pkg{parallel} package. (Default:
\code{options("utiml.cores", 1)})}

\item{seed}{An optional integer used to set the seed. This is useful when
the method is run in parallel. (Default: \code{options("utiml.seed", NA)})}
}
\value{
An object of class \code{EBRmodel} containing the set of fitted
  BR models, including:
\describe{
  \item{models}{A list of BR models.}
  \item{nrow}{The number of instances used in each training dataset.}
  \item{ncol}{The number of attributes used in each training dataset.}
  \item{rounds}{The number of interations.}
}
}
\description{
Create an Ensemble of Binary Relevance model for multilabel classification.
}
\details{
This model is composed by a set of Binary Relevance models. Binary Relevance
is a simple and effective transformation method to predict multi-label data.
}
\note{
If you want to reproduce the same classification and obtain the same
 result will be necessary set a flag utiml.mc.set.seed to FALSE.
}
\examples{
model <- ebr(toyml, "RANDOM")
pred <- predict(model, toyml)

\dontrun{
# Use J48 with 90\% of instances and only 5 rounds
model <- ebr(toyml, 'J48', m = 5, subsample = 0.9)

# Use 75\% of attributes
model <- ebr(dataset$train, attr.space = 0.75)

# Running in 4 cores and define a specific seed
model1 <- ebr(toyml, cores=4, seed = 312)
}
}
\references{
Read, J., Pfahringer, B., Holmes, G., & Frank, E. (2011). Classifier
   chains for multi-label classification. Machine Learning, 85(3), 333-359.

   Read, J., Pfahringer, B., Holmes, G., & Frank, E. (2009).
   Classifier Chains for Multi-label Classification. Machine Learning and
   Knowledge Discovery in Databases, Lecture Notes in Computer Science,
   5782, 254-269.
}
\seealso{
Other Ensemble methods: \code{\link{ecc}}

Other Transformation methods: \code{\link{brplus}},
  \code{\link{br}}, \code{\link{cc}}, \code{\link{clr}},
  \code{\link{ctrl}}, \code{\link{dbr}}, \code{\link{ecc}},
  \code{\link{lift}}, \code{\link{lp}}, \code{\link{mbr}},
  \code{\link{ns}}, \code{\link{prudent}},
  \code{\link{rakel}}, \code{\link{rdbr}},
  \code{\link{rpc}}
}

