% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecc.R
\name{ecc}
\alias{ecc}
\title{Ensemble of Classifier Chains for multi-label Classification}
\usage{
ecc(mdata, base.method = "SVM", m = 10, subsample = 0.75,
  attr.space = 0.5, ..., CORES = 1)
}
\arguments{
\item{mdata}{Object of class \code{\link[mldr]{mldr}}, a multi-label train
dataset (provided by \pkg{mldr} package).}

\item{base.method}{A string with the name of base method. The same base method
  will be used for train all subproblems.

  Default valid options are: \code{'SVM'}, \code{'C4.5'}, \code{'C5.0'},
  \code{'RF'}, \code{'NB'} and \code{'KNN'}. To use other base method see
  \code{\link{mltrain}} and \code{\link{mlpredict}} instructions. (default:
   \code{'SVM'})}

\item{m}{The number of Binary Relevance models used in the ensemble.}

\item{subsample}{A value between 0.1 and 1 to determine the percentage of
training instances must be used for each interation. (default: 0.75)}

\item{attr.space}{A value between 0.1 and 1 to determine the percentage of
attributes must be used for each interation. (default: 0.50)}

\item{...}{Others arguments passed to the base method for all subproblems.}

\item{CORES}{The number of cores to parallelize the training. Values higher
than 1 require the \pkg{parallel} package. (default: 1)}
}
\value{
An object of class \code{ECCmodel} containing the set of fitted
  CC models, including: \describe{
  \item{rounds}{The number of interations}
  \item{models}{A list of BR models.}
  \item{nrow}{The number of instances used in each training dataset}
  \item{ncol}{The number of attributes used in each training dataset}
}
}
\description{
Create an Ensemble of Classifier Chains model for
  multilabel classification.

  This model is composed by a set of Classifier Chains models.
  Classifier Chains is a Binary Relevance transformation method based
  to predict multi-label data. It is different from BR method due the strategy
  of extended the attribute space with the 0/1 label relevances of all previous
  classifiers, forming a classifier chain.
}
\examples{
# Train and predict emotion multilabel dataset using Ensemble of Classifier Chains
dataset <- mldr_random_holdout(emotions, c(train=0.9, test=0.1))

# Use all default values
model <- ecc(dataset$train)
pred <- predict(model, dataset$test)

# Use C4.5 with 100\% of instances and only 5 rounds
model <- ecc(dataset$train, "C4.5", m = 5, subsample = 1)
pred <- predict(model, dataset$test)

# Use 75\% of attributes
model <- ecc(dataset$train, attr.space = 0.75)
pred <- predict(model, dataset$test)

# Running in 4 cores
model <- ecc(dataset$train, CORES=4)
pred <- predict(model, dataset$test, CORES=4)
}
\references{
Read, J., Pfahringer, B., Holmes, G., & Frank, E. (2011). Classifier
   chains for multi-label classification. Machine Learning, 85(3), 333–359.

   Read, J., Pfahringer, B., Holmes, G., & Frank, E. (2009).
   Classifier Chains for Multi-label Classification. Machine Learning and
   Knowledge Discovery in Databases, Lecture Notes in Computer Science,
   5782, 254–269.
}
\seealso{
\code{\link[=ecc]{Ensemble of classifier Chains (ECC)}}

Other Ensemble methods: \code{\link{ctrl}},
  \code{\link{ebr}}

Other Transformation methods: \code{\link{brplus}},
  \code{\link{br}}, \code{\link{cc}}, \code{\link{ctrl}},
  \code{\link{dbr}}, \code{\link{ebr}}, \code{\link{mbr}},
  \code{\link{ns}}, \code{\link{prudent}},
  \code{\link{rdbr}}
}

