% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threshold.R
\name{fixed_threshold}
\alias{fixed_threshold}
\title{Apply a fixed threshold in the results}
\usage{
fixed_threshold(prediction, threshold = 0.5)
}
\arguments{
\item{prediction}{A matrix with scores/probabilities where the columns
are the labels and the rows are the instances.}

\item{threshold}{A single value between 0 and 1 or a list with threshold
values contained one value per label.}
}
\value{
A matrix with bipartition results
}
\description{
Transfom a prediction matrix with scores/probabilities in a bipartion
prediction matrix. A global fixed threshold can be used of all labels or
different fixed thresholds, one for each label.
}
\examples{
# Create a prediction matrix with scores
result <- matrix(
 data = rnorm(9, 0.5, 0.2),
 ncol = 3,
 dimnames = list(NULL, c('lbl1',  'lb2', 'lb3'))
)

# Use 0.5 as threshold
fixed_threshold(result)

# Use an threshold for each label
fixed_threshold(result, c(0.4, 0.6, 0.7))
}
\references{
Al-Otaibi, R., Flach, P., & Kull, M. (2014). Multi-label Classification: A
 Comparative Study on Threshold Selection Methods. In First International
 Workshop on Learning over Multiple Contexts (LMCE) at ECML-PKDD 2014.
}
\seealso{
Other threshold: \code{\link{mcut_threshold}},
  \code{\link{mcut_threshold.default}},
  \code{\link{mcut_threshold.mlresult}};
  \code{\link{pcut_threshold}},
  \code{\link{pcut_threshold.default}},
  \code{\link{pcut_threshold.mlresult}};
  \code{\link{rcut_threshold}},
  \code{\link{rcut_threshold.default}},
  \code{\link{rcut_threshold.mlresult}};
  \code{\link{subset_correction}}
}

