% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlresult.R
\name{is.bipartition}
\alias{is.bipartition}
\title{Test if a mlresult contains crisp values as default}
\usage{
is.bipartition(mlresult)
}
\arguments{
\item{mlresult}{The mlresult object}
}
\value{
logical value
}
\description{
Test if a mlresult contains crisp values as default
}

