% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mbr.R
\name{mbr}
\alias{mbr}
\title{Meta-BR or 2BR for multi-label Classification}
\usage{
mbr(mdata, base.method = "SVM", phi = 0, ..., predict.params = list(),
  save.datasets = FALSE, CORES = 1)
}
\arguments{
\item{mdata}{Object of class \code{\link[mldr]{mldr}}, a multi-label train
dataset (provided by \pkg{mldr} package).}

\item{base.method}{A string with the name of base method. The same base method
  will be used for train all subproblems in the two rounds. This is slightly
  different from the original paper that uses two distinct values from this
  parameter.

  Default valid options are: \code{'SVM'}, \code{'C4.5'}, \code{'C5.0'},
  \code{'RF'}, \code{'NB'} and \code{'KNN'}. To use other base method see
  \code{\link{mltrain}} and \code{\link{mlpredict}} instructions. (default:
   \code{'SVM'})}

\item{phi}{A value between 0 and 1 to determine the correlation coefficient,
The value 0 include all labels in the second phase and the 1 only the
predicted label. A good value for this argument is 0.3 as suggest in
original paper. (default: 0)}

\item{...}{Others arguments passed to the base method for all subproblems.}

\item{predict.params}{A list of default arguments passed to the predictor a mldr object
method. (default: \code{list()})}

\item{save.datasets}{Logical indicating whether the binary datasets must be
saved in the model or not. (default: \code{FALSE})}

\item{CORES}{The number of cores to parallelize the training. Values higher
  than 1 require the \pkg{parallel} package. (default: 1)}
}
\value{
An object of class \code{MBRmodel} containing the set of fitted
  models, including: \describe{
     \item{labels}{A vector with the label names.}
     \item{phi}{The value of \code{phi} parameter.}
     \item{correlation}{The matrix of label correlations used in combination
       with \code{phi} parameter to define the labels used in the second step.
     }
     \item{basemodel}{The BRModel used in the first iteration.}
     \item{metamodels}{A list of models named by the label names used in the
       second iteration.
     }
     \item{datasets}{A list with \code{base} and \code{meta} datasets of
       type \code{mldBR} named by the label names. Only when the
       \code{save.datasets = TRUE}.
     }
 }
}
\description{
Create a Meta-BR (MBR) classifier to predic multi-label
 data. To this, two round of Binary Relevance is executed, such that,
 the first iteraction generates new attributes to enrich the second
 prediction.

 This implementation use complete training set for both training and
 prediction steps of 2BR. However, the \code{phi} parameter may be used
 to remove low label correlations on the second step. Furthermore, we
 remove the \code{F} parameter because its specification is not clear.
}
\examples{
# Train and predict emotion multilabel dataset using Meta-BR
library(utiml)
testdata <- emotions$dataset[sample(1:100, 10), emotions$attributesIndexes]

# Use SVM as base method
model <- mbr(emotions)
pred <- predict(model, testdata)

# Use different phi correlation with C4.5 classifier
model <- mbr(emotions, "C4.5", 0.3)
pred <- predict(model, testdata)

# Set a specific parameter
model <- mbr(emotions, "KNN", k=5)
pred <- predict(model, testdata)
}
\references{
Tsoumakas, G., Dimou, A., Spyromitros, E., Mezaris, V., Kompatsiaris, I., &
   Vlahavas, I. (2009). Correlation-based pruning of stacked binary relevance models
   for multi-label learning. In Proceedings of the Workshop on Learning from
   Multi-Label Data (MLD’09) (pp. 22–30).
 Godbole, S., & Sarawagi, S. (2004). Discriminative Methods for Multi-labeled
   Classification. In Data Mining and Knowledge Discovery (pp. 1–26).
}
\seealso{
\code{\link{labels_correlation_coefficient}}

Other Transformation methods: \code{\link{brplus}};
  \code{\link{br}}; \code{\link{cc}}; \code{\link{dbr}};
  \code{\link{ebr}}; \code{\link{ecc}}; \code{\link{ns}}
}

