% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{mlmeasure_micro_accuracy}
\alias{mlmeasure_micro_accuracy}
\title{Multi-label Micro-Accuracy Measure}
\usage{
mlmeasure_micro_accuracy(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}
}
\description{
Multi-label Micro-Accuracy Measure
}
\references{
Gibaja, E., & Ventura, S. (2015). A Tutorial on Multilabel
 Learning. ACM Comput. Surv., 47(3), 52:1–52:38.
}

