% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_br.R
\name{predict.BRmodel}
\alias{predict.BRmodel}
\title{Predict Method for Binary Relevance}
\usage{
\method{predict}{BRmodel}(object, newdata,
  probability = getOption("utiml.use.probs", TRUE), ...,
  cores = getOption("utiml.cores", 1), seed = getOption("utiml.seed", NA))
}
\arguments{
\item{object}{Object of class '\code{BRmodel}'.}

\item{newdata}{An object containing the new input data. This must be a
matrix, data.frame or a mldr object.}

\item{probability}{Logical indicating whether class probabilities should be
returned. (Default: \code{getOption("utiml.use.probs", TRUE)})}

\item{...}{Others arguments passed to the base algorithm prediction for all
subproblems.}

\item{cores}{The number of cores to parallelize the training. Values higher
than 1 require the \pkg{parallel} package. (Default:
\code{options("utiml.cores", 1)})}

\item{seed}{An optional integer used to set the seed. This is useful when
the method is run in parallel. (Default: \code{options("utiml.seed", NA)})}
}
\value{
An object of type mlresult, based on the parameter probability.
}
\description{
This function predicts values based upon a model trained by \code{\link{br}}.
}
\examples{
model <- br(toyml, "RANDOM")
pred <- predict(model, toyml)

\dontrun{
# Predict SVM scores
model <- br(toyml, "SVM")
pred <- predict(model, toyml)

# Predict SVM bipartitions running in 4 cores
pred <- predict(model, toyml, probability = FALSE, CORES = 4)

# Passing a specif parameter for SVM predict algorithm
pred <- predict(model, dataset$test, na.action = na.fail)
}
}
\seealso{
\code{\link[=br]{Binary Relevance (BR)}}
}
