% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{predict.CCmodel}
\alias{predict.CCmodel}
\title{Predict Method for Classifier Chains}
\usage{
\method{predict}{CCmodel}(object, newdata, probability = TRUE, ...)
}
\arguments{
\item{object}{Object of class "\code{CCmodel}", created by \code{\link{cc}} method.}

\item{newdata}{An object containing the new input data. This must be a matrix or
data.frame object containing the same size of training data or a mldr object.}

\item{probability}{Logical indicating whether class probabilities should be returned.
(default: \code{TRUE})}

\item{...}{Others arguments passed to the base method prediction for all
subproblems.}
}
\value{
A matrix containing the probabilistic values or just predictions (only when
  \code{probability = FALSE}). The rows indicate the predicted object and the
  columns indicate the labels.
}
\description{
This function predicts values based upon a model trained by \code{cc}.
}
\examples{
# Emotion multi-label dataset using Classifier Chains
dataset <- mldr_random_holdout(emotions, c(train=0.9, test=0.1))

# Predict SVM scores
model <- cc(dataset$train)
pred <- predict(model, dataset$test)

# Predict SVM bipartitions
pred <- predict(model, dataset$test, prob = FALSE)

# Passing a specif parameter for SVM predict method
pred <- predict(model, dataset$test, na.action = na.fail)
}
\seealso{
\code{\link[=cc]{Classifier Chains (CC)}}
}

