% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrl.R
\name{predict.CTRLmodel}
\alias{predict.CTRLmodel}
\title{Predict Method for CTRL}
\usage{
\method{predict}{CTRLmodel}(object, newdata, vote.schema = "maj",
  probability = TRUE, ..., CORES = 1)
}
\arguments{
\item{object}{Object of class '\code{CTRLmodel}', created by \code{\link{ctrl}} method.}

\item{newdata}{An object containing the new input data. This must be a matrix or
data.frame object containing the same size of training data or a mldr object.}

\item{probability}{Logical indicating whether class probabilities should be returned.
(default: \code{TRUE})}

\item{...}{Others arguments passed to the base method prediction for all
subproblems.}

\item{CORES}{The number of cores to parallelize the prediction. Values higher
than 1 require the \pkg{parallel} package (default: 1).}
}
\value{
A matrix containing the probabilistic values or just predictions (only when
  \code{probability = FALSE}). The rows indicate the predicted object and the
  columns indicate the labels.
}
\description{
This function predicts values based upon a model trained by
 \code{\link{ctrl}}.
}
\examples{
# Emotion multi-label dataset using Binary Relevance
dataset <- mldr_random_holdout(emotions, c(train=0.9, test=0.1))

# Predict SVM scores
model <- ctrl(dataset$train)
pred <- predict(model, dataset$test)

# Predict SVM bipartitions running in 6 cores
pred <- predict(model, dataset$test, probability = FALSE, CORES = 6)

# Using the Maximum vote schema
pred <- predict(model, dataset$test, vote.schema = 'max')
}
\seealso{
\code{\link[=ctrl]{CTRL}}
}

