% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbr.R
\name{predict.DBRmodel}
\alias{predict.DBRmodel}
\title{Predict Method for DBR}
\usage{
\method{predict}{DBRmodel}(object, newdata, ..., estimative = NULL,
  probability = TRUE, CORES = 1)
}
\arguments{
\item{object}{Object of class "\code{DBRmodel}", created by \code{\link{dbr}} method.}

\item{newdata}{An object containing the new input data. This must be a matrix or
data.frame object containing the same size of training data or a mldr object.}

\item{...}{Others arguments passed to the base method prediction for all
subproblems.}

\item{estimative}{A matrix containing the result of other multi-label classification
algorithm. This table must contain only 0 or 1 predictions and it must be a
multi-label prediction result.}

\item{probability}{Logical indicating whether class probabilities should be returned.
(default: \code{TRUE})}

\item{CORES}{The number of cores to parallelize the prediction. Values higher
  than 1 require the \pkg{parallel} package (default: 1).}
}
\value{
A matrix containing the probabilistic values or just predictions (only when
  \code{probability = FALSE}). The rows indicate the predicted object and the
  columns indicate the labels.
}
\description{
This function predicts values based upon a model trained by \code{dbr}.
In general this method is a restricted version of \code{\link{predict.BRPmodel}}
using the 'NU' strategy.

As new feature is possible to use other multi-label classifier to predict the estimate
values of each label. To this use the prediction argument to inform a result of other
multi-label algorithm.
}
\examples{
#' library(utiml)

# Emotion multi-label dataset using DBR
testdata <- emotions$dataset[sample(1:100, 10), emotions$attributesIndexes]

# Predict SVM scores
model <- dbr(emotions)
pred <- predict(model, testdata)

# Passing a specif parameter for SVM predict method
pred <- predict(model, testdata, na.action = na.fail)

# Using other classifier (EBR) to made the labels estimatives
estimative <- predict(ebr(emotions), testdata, probability = FALSE)
model <- dbr(emotions, estimate.models = FALSE)
pred <- predict(model, testdata, estimative = estimative)
}
\references{
Montañes, E., Senge, R., Barranquero, J., Ramón Quevedo, J., José Del Coz,
   J., & Hüllermeier, E. (2014). Dependent binary relevance models for
   multi-label classification. Pattern Recognition, 47(3), 1494–1508.
}
\seealso{
\code{\link[=dbr]{Dependent Binary Relevance (DBR)}}
}

