% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_mbr.R
\name{predict.MBRmodel}
\alias{predict.MBRmodel}
\title{Predict Method for Meta-BR/2BR}
\usage{
\method{predict}{MBRmodel}(object, newdata, probability = TRUE, ...,
  CORES = 1)
}
\arguments{
\item{object}{Object of class '\code{MBRmodel}', created by \code{\link{mbr}} method.}

\item{newdata}{An object containing the new input data. This must be a matrix or
data.frame object containing the same size of training data or a mldr object.}

\item{probability}{Logical indicating whether class probabilities should be returned.
(default: \code{TRUE})}

\item{...}{Others arguments passed to the base method prediction for all
subproblems.}

\item{CORES}{The number of cores to parallelize the prediction. Values higher
  than 1 require the \pkg{parallel} package (default: 1).}
}
\value{
A matrix containing the probabilistic values or just predictions (only when
  \code{probability = FALSE}). The rows indicate the predicted object and the
  columns indicate the labels.
}
\description{
This function predicts values based upon a model trained by \code{mbr}.
}
\examples{
# Emotion multi-label dataset using Meta-BR or 2BR
dataset <- mldr_random_holdout(emotions, c(train=0.9, test=0.1))

# Predict SVM scores
model <- mbr(dataset$train)
pred <- predict(model, dataset$test)

# Predict SVM bipartitions
pred <- predict(model, dataset$test, probability = FALSE)

# Passing a specif parameter for SVM predict method
pred <- predict(model, dataset$test, na.action = na.fail)
}
\seealso{
\code{\link[=mbr]{Meta-BR (MBR or 2BR)}}
}

