% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prudent.R
\name{predict.PruDentmodel}
\alias{predict.PruDentmodel}
\title{Predict Method for PruDent}
\usage{
\method{predict}{PruDentmodel}(object, newdata, ..., probability = TRUE,
  CORES = 1)
}
\arguments{
\item{object}{Object of class "\code{PruDentmodel}", created by \code{\link{prudent}} method.}

\item{newdata}{An object containing the new input data. This must be a matrix or
data.frame object containing the same size of training data or a mldr object.}

\item{...}{Others arguments passed to the base method prediction for all
subproblems.}

\item{probability}{Logical indicating whether class probabilities should be returned.
(default: \code{TRUE})}

\item{CORES}{The number of cores to parallelize the prediction. Values higher
  than 1 require the \pkg{parallel} package (default: 1).}
}
\value{
A matrix containing the probabilistic values or just predictions (only when
  \code{probability = FALSE}). The rows indicate the predicted object and the
  columns indicate the labels.
}
\description{
This function predicts values based upon a model trained by \code{prudent}.
}
\examples{
#' library(utiml)

# Emotion multi-label dataset using PruDent
testdata <- emotions$dataset[sample(1:100, 10), emotions$attributesIndexes]

# Predict SVM scores
model <- prudent(emotions)
pred <- predict(model, testdata)

# Predict SVM bipartitions
pred <- predict(model, testdata, probability = FALSE)

# Passing a specif parameter for SVM predict method
pred <- predict(model, testdata, na.action = na.fail)
}
\seealso{
\code{\link[=prudent]{PruDent}}
}

