% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_lift.R
\name{print.LIFTmodel}
\alias{print.LIFTmodel}
\title{Print LIFT model}
\usage{
\method{print}{LIFTmodel}(x, ...)
}
\arguments{
\item{x}{The lift model}

\item{...}{ignored}
}
\description{
Print LIFT model
}
