% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_rakel.R
\name{print.RAkELmodel}
\alias{print.RAkELmodel}
\title{Print RAkEL model}
\usage{
\method{print}{RAkELmodel}(x, ...)
}
\arguments{
\item{x}{The rakel model}

\item{...}{ignored}
}
\description{
Print RAkEL model
}
