% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{print.randomModel}
\alias{print.randomModel}
\title{Print Random model}
\usage{
\method{print}{randomModel}(x, ...)
}
\arguments{
\item{x}{The base model}

\item{...}{ignored}
}
\description{
Print Random model
}

