% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_prudent.R
\name{prudent}
\alias{prudent}
\title{PruDent classifier for multi-label Classification}
\usage{
prudent(mdata, base.method = getOption("utiml.base.method", "SVM"), phi = 0,
  ..., CORES = getOption("utiml.cores", 1))
}
\arguments{
\item{mdata}{A mldr dataset used to train the binary models.}

\item{base.method}{A string with the name of the base method. (Default:
\code{options("utiml.base.method", "SVM")})}

\item{phi}{A value between 0 and 1 to determine the information gain. The
value 0 include all labels in the second phase and the 1 none.}

\item{...}{Others arguments passed to the base method for all subproblems.}

\item{CORES}{The number of cores to parallelize the training. Values higher
than 1 require the \pkg{parallel} package. (Default:
\code{options("utiml.cores", 1)})}
}
\value{
An object of class \code{PruDentmodel} containing the set of fitted
  models, including:
  \describe{
   \item{labels}{A vector with the label names.}
   \item{phi}{The value of \code{phi} parameter.}
   \item{IG}{The matrix of Information Gain used in combination
     with \code{phi} parameter to define the labels used in the second step.
   }
   \item{basemodel}{The BRModel used in the first iteration.}
   \item{metamodels}{A list of models named by the label names used in the
     second iteration.
   }
}
}
\description{
Create a PruDent (MBR) classifier to predic multi-label data. To this, two
round of Binary Relevance is executed, such that, the first iteraction
generates new attributes to enrich the second prediction.
}
\details{
In the second phase only labels whose information gain is greater than a
specific phi value is added.
}
\examples{
\dontrun{
# Use SVM as base method
model <- prudent(toyml)
pred <- predict(model, toyml)

# Use different phi correlation with J48 classifier
model <- prudent(toyml, 'J48', 0.3)

# Set a specific parameter
model <- prudent(toyml, 'KNN', k=5)
}
}
\references{
Alali, A., & Kubat, M. (2015). PruDent: A Pruned and Confident Stacking
   Approach for Multi-Label Classification. IEEE Transactions on Knowledge
   and Data Engineering, 27(9), 2480–2493.
}
\seealso{
\code{\link{calculate_labels_information_gain}}

Other Transformation methods: \code{\link{brplus}},
  \code{\link{br}}, \code{\link{cc}}, \code{\link{ctrl}},
  \code{\link{dbr}}, \code{\link{ebr}}, \code{\link{ecc}},
  \code{\link{mbr}}, \code{\link{ns}}, \code{\link{rdbr}}
}

