% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{summary.mltransformation}
\alias{summary.mltransformation}
\title{Summary method for mltransformation}
\usage{
\method{summary}{mltransformation}(object, ...)
}
\arguments{
\item{object}{A transformed dataset}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
No return value, called for print model's detail
}
\description{
Summary method for mltransformation
}
