% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{utiml_compute_multilabel_ensemble}
\alias{utiml_compute_multilabel_ensemble}
\title{Compute the ensemble predictions based on some vote schema}
\usage{
utiml_compute_multilabel_ensemble(predictions, vote.schema,
  probability = TRUE)
}
\arguments{
\item{predictions}{A list of mlresult}

\item{vote.schema}{Define the way that ensemble must compute the predictions.
The valid options are describe in \link{utiml_vote.schema_method}.}

\item{probability}{A logical value. If \code{TRUE} the predicted values are
the score between 0 and 1, otherwise the values are bipartition 0 or 1.}
}
\value{
A new mlresult that final result
}
\examples{
predictions <- list()
predictions$model1 <- prediction(brmodel1, testdata)
predictions$model2 <- prediction(brmodel2, testdata)
result <- utiml_compute_ensemble_predictions(predictions, 'MAJ')
}

