% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampling.R
\name{utiml_holdout}
\alias{utiml_holdout}
\title{Create a holdout partition based in the specified method}
\usage{
utiml_holdout(mdata, partitions, partition.names, holdout.method)
}
\arguments{
\item{mdata}{A dataset of class \code{\link[mldr]{mldr}}.}

\item{partitions}{A list of percentages with partitions sizes.}

\item{partition.names}{a vector with the partition names.}

\item{holdout.method}{The method to split the data.}
}
\value{
A list with at least two datasets sampled as specified
 in partitions parameter.
}
\description{
Create a holdout partition based in the specified method
}
\examples{
utiml_holdout(mdata, partitions, partition.names, utiml_random_split)
}
\seealso{
Other sampling: \code{\link{mldr_getfold}};
  \code{\link{mldr_iterative_stratification_holdout}};
  \code{\link{mldr_iterative_stratification_kfold}};
  \code{\link{mldr_random_holdout}};
  \code{\link{mldr_random_kfold}};
  \code{\link{mldr_stratified_holdout}};
  \code{\link{mldr_stratified_kfold}};
  \code{\link{utiml_kfold}}
}

