% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinternal.R
\name{utiml_ifelse}
\alias{utiml_ifelse}
\title{Conditional value selection}
\usage{
utiml_ifelse(test, yes, no)
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}

\item{yes}{object that will be returned when the test value is true.}

\item{no}{object that will be returned when the test value is false}
}
\value{
The respective value yes or no based on test value. This
 is an alternative way to use a single logical value for avoid
 the real if/else for choice lists, matrices and other composed data.
}
\description{
Conditional value selection
}
\examples{
utiml_ifelse(TRUE, dataframe1, dataframe2) #dataframe1
utiml_ifelse(length(my.list) > 10, my.list[1:10], my.list)
}
\seealso{
\code{\link{ifelse}}
}

