% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_lapply}
\alias{utiml_lapply}
\title{Select the suitable method lapply or mclaplly}
\usage{
utiml_lapply(mylist, myfnc, cores, ...)
}
\arguments{
\item{mylist}{a list to iterate.}

\item{myfnc}{The function to be applied to each element of the mylist.}

\item{cores}{The number of cores to use. If 1 use lapply oterwise use
mclapply.}

\item{...}{Extra arguments to myfnc.}
}
\value{
A list with the results of the specified method.
}
\description{
Select the suitable method lapply or mclaplly
}
\examples{
\dontrun{
utiml_lapply(c(4,9,27), sqrt, 1) #use lapply
utiml_lapply(c(4,9,27), sqrt, 3) #use mclapply
}
}

