% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{utiml_measure_precision}
\alias{utiml_measure_precision}
\title{Multi-label Precision Measure}
\usage{
utiml_measure_precision(mlconfmat, ...)
}
\arguments{
\item{mlconfmat}{Confusion matrix}

\item{...}{ignored}
}
\description{
Multi-label Precision Measure
}
\references{
Godbole, S., & Sarawagi, S. (2004). Discriminative Methods for
Multi-labeled Classification. In Proceedings of the 8th Pacific-Asia
Conference on Knowledge Discovery and Data Mining (PAKDD 2004) (pp. 22-30).
}

