% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{utiml_newdata}
\alias{utiml_newdata}
\alias{utiml_newdata.default}
\alias{utiml_newdata.mldr}
\title{Return the newdata to a data.frame or matrix}
\usage{
utiml_newdata(newdata)

\method{utiml_newdata}{default}(newdata)

\method{utiml_newdata}{mldr}(newdata)
}
\arguments{
\item{newdata}{The data.frame or mldr data}
}
\value{
A dataframe or matrix containing only dataset
}
\description{
Return the newdata to a data.frame or matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{mldr}: 
}}
\examples{
test <- emotions$dataset[,emotions$attributesIndexes]
all(test == utiml_newdata(emotions)) # TRUE
all(test == utiml_newdata(test)) # TRUE
}

